Split background-agent-runner.ts into modules by pure moves
The 10,141-line runner file is split into eight modules under
runtime/src/app-server/background-agent-runner/. Every moved declaration
keeps its exact text, leading comments included; only import and export
statements were added. background-agent-runner.ts keeps the
AgenCDelegateBackgroundAgentRunner class, the turn-driver ingress
(prepareDaemonUserPrompt, installDaemonTurnDriverHooks), eventFromProgress
and messageText, and re-exports every name it exported before, so no
importer or test changes.

Where each group went:
- shared.ts: the exported params, results, snapshot and runner interface
  types, the two error classes, the DAEMON_USER_PROMPT_PREPARED symbol,
  the internal ActiveBackgroundAgent and daemon event types, and the
  small pure helpers (sequence guards, JSON guards, hashStable,
  canonicalEventId, historyEpochForBoundary, messageContentFingerprint).
- snapshot-retention.ts: bounded event buffers and gap markers,
  submission and shell caches, usage and cost snapshots, managedTokenUsage.
- progress-events.ts: phaseEventToProgressEvent,
  canonicalSessionEventFromRecoveredProgress, interruptedToolResultContent.
- daemon-events.ts: daemonEventFromUnboundSessionEvent, correlateDaemonEvent,
  notificationFromDaemonEvent and the run status mappings.
- journal-reconstruction.ts: sessionTranscriptV2FromRollout, transcript
  boundaries, findPersistedMessageSubmission, currentRunEpochFromRollout.
- turn-lifecycle.ts: active-agent predicates, shell submission helpers,
  durable run startup, cancellation, terminal and suspension events.
- tool-recovery.ts: recovered tool-call replay, restoredAgentMetadata,
  planApprovalPayloadFields and the permission decision helpers.
- runtime-settings.ts: durable runtime settings, the permission authority
  coordinator, restore overrides, buildBootstrapArgv and the unattended
  policy installer.

One declaration needed a one-token edit: phaseEventToProgressEvent types
its parameter with an inline import("../phases/events.js"), which now
lives one directory deeper and reads import("../../phases/events.js").

knip.config.mjs lists the new module files next to background-agent-runner.ts
in appServerContractExportFiles, so check:unused reports the same findings
as before for the test-facing exports managedTokenUsage and
planApprovalPayloadFields.
