+++ runtime/knip.config.mjs
@@ -375,6 +375,14 @@ const appServerContractExportFiles = [
   "src/app-server/agent-cli.ts",
   "src/app-server/auth.ts",
   "src/app-server/background-agent-runner.ts",
+  "src/app-server/background-agent-runner/daemon-events.ts",
+  "src/app-server/background-agent-runner/journal-reconstruction.ts",
+  "src/app-server/background-agent-runner/progress-events.ts",
+  "src/app-server/background-agent-runner/runtime-settings.ts",
+  "src/app-server/background-agent-runner/shared.ts",
+  "src/app-server/background-agent-runner/snapshot-retention.ts",
+  "src/app-server/background-agent-runner/tool-recovery.ts",
+  "src/app-server/background-agent-runner/turn-lifecycle.ts",
   "src/app-server/client-multiplexer.ts",
   "src/app-server/daemon-autostart.ts",
   "src/app-server/daemon-cli.ts",
