+++ runtime/src/app-server/background-agent-runner.ts
@@ -7,84 +7,46 @@
  * response is returned.
  */
 
-import { createHash, randomUUID } from "node:crypto";
+import { randomUUID } from "node:crypto";
 import { readFileSync, readdirSync } from "node:fs";
 import { join as joinPath } from "node:path";
-
 import { roughTokenCountEstimation } from "../llm/token-estimation.js";
 import {
   bootstrapLocalRuntimeSession,
-  type BootstrapLocalRuntimeSessionOptions,
   type LocalRuntimeBootstrap,
-  type PreparedConfiguredExecutionAuthority,
 } from "../bin/bootstrap.js";
-import { buildStructuredSessionBootstrapArgv } from "./session-bootstrap-argv.js";
 import { ensureAgentControl } from "../bin/delegate-tool.js";
 import { clearSession } from "../commands/clear.js";
-import type { AgentControl } from "../agents/control.js";
-import { MailboxClosedError } from "../agents/mailbox.js";
 import { runTurn } from "../session/run-turn.js";
 import {
   prepareUserPromptForTurn,
   userPromptDisplayText,
 } from "../hooks/user-prompt-ingress.js";
-import {
-  ROOT_AGENT_PATH,
-  joinAgentPath,
-  normalizeAgentMetadata,
-  normalizeAgentNameForPath,
-  type AgentMetadata,
-  type AgentPath,
-} from "../agents/registry.js";
-import type { AgentThread } from "../agents/thread.js";
+import type { AgentPath } from "../agents/registry.js";
 import type { ManagedThread } from "../agents/thread-manager.js";
 import { ConversationThreadManager } from "../conversation/thread-manager.js";
-import {
-  runAgent,
-  type RunAgentProgressEvent,
-  type RunAgentResult,
-} from "../agents/run-agent.js";
+import type { RunAgentProgressEvent } from "../agents/run-agent.js";
 import type { AuthBackend } from "../auth/backend.js";
 import type { LLMContentPart, LLMMessage } from "../llm/types.js";
-import { freshDenialTracking } from "../permissions/denial-tracking.js";
-import {
-  attachContextDefaults,
-  hasPermissionsToUseTool,
-  type AppStateSnapshot,
-  type ToolEvaluatorContext,
-} from "../permissions/evaluator.js";
 import type { ApprovalCtx, ApprovalResolver } from "../tools/orchestrator.js";
 import { routerFromRegistry } from "../tools/router.js";
 import { buildLiveToolDispatchOptions } from "../phases/execute-tools.js";
-import type { ToolRecoveryCategory } from "../tools/types.js";
-import {
-  classifyUntrustedToolResult,
-  frameUntrustedToolResultContent,
-} from "../tools/untrusted-tool-result-framing.js";
 import type { ToolDispatchResult, ToolRegistry } from "../tool-registry.js";
-import { getPlan, getPlanFilePath } from "../utils/plans.js";
 import { stableStringify } from "../utils/stableStringify.js";
 import { logForDebugging } from "../utils/debug.js";
-import { EXIT_PLAN_MODE_TOOL_NAME } from "../tools/ExitPlanModeTool/constants.js";
-import type { AgentId } from "../types/ids.js";
-import {
-  computeUsdCost,
-  DEFAULT_MODEL_COSTS,
-  type ModelUsage,
-} from "../session/cost.js";
 import {
   runWithBootstrapSessionScope,
   runWithCurrentRuntimeSession,
 } from "../session/current-session.js";
-import { runWithCanonicalSettingsAuthority } from "../utils/settings/canonicalAuthority.js";
+import {
+  runWithCanonicalSettingsAuthority,
+} from "../utils/settings/canonicalAuthority.js";
 import { resolveDefaultShell } from "../utils/shell/resolveDefaultShell.js";
 import { escapeXml } from "../utils/xml.js";
 import {
   canCycleToAuto,
   createDisabledAutoModeContext,
   transitionPermissionMode,
-  type PermissionContextPublication,
-  type PermissionModeRegistry,
 } from "../permissions/permission-mode.js";
 import {
   applyPermissionRulesSnapshot,
@@ -107,21 +69,18 @@ import {
   PermissionRuleMutationPrecommitError,
 } from "../permissions/permission-updates.js";
 import { applyModelSwitch } from "../commands/model.js";
-import type { ProviderModelSelectionOutcome } from "../contracts/provider-model-selection.js";
+import type {
+  ProviderModelSelectionOutcome,
+} from "../contracts/provider-model-selection.js";
 import {
   readSessionSelection,
   resolveProviderModelSelection,
 } from "../session/provider-model-selection.js";
 import { applyProviderSwitch } from "../commands/provider.js";
 import { resolveProfile } from "../config/profiles.js";
-import { mergeProviderModelLayer } from "../config/provider-model-authority.js";
-import type { AgenCConfig } from "../config/schema.js";
 import {
-  COORDINATED_CONFIG_STORE_PUBLICATION,
-  type PreparedConfigStoreReload,
-} from "../config/store.js";
-import type { McpRefreshResult } from "../session/mcp-startup.js";
-import { resolveLiveEffectPoison } from "../budget/effect-settlement-supervisor.js";
+  resolveLiveEffectPoison,
+} from "../budget/effect-settlement-supervisor.js";
 import {
   resolveLiveDurableEffectReview,
   type ResolveDurableEffectReviewOptions,
@@ -129,16 +88,15 @@ import {
 } from "../state/effect-review.js";
 import { openStateDatabases } from "../state/sqlite-driver.js";
 import { mergeDaemonClientEnvironment } from "./client-env-snapshot.js";
-
-import { permissionGrantsFromToolPermissionContext } from "../permissions/permission-grants.js";
-import { applyUnattendedPermissionPolicyToContext } from "../permissions/unattended-policy.js";
+import {
+  permissionGrantsFromToolPermissionContext,
+} from "../permissions/permission-grants.js";
 import {
   ABORT,
   DENIED,
   TIMED_OUT,
   type ReviewDecision,
 } from "../permissions/review-decision.js";
-import type { AgentStatus as ThreadAgentStatus } from "../agents/status.js";
 import type {
   McpServerMutationResult,
   McpSurfaceSnapshot,
@@ -146,29 +104,18 @@ import type {
   Session,
 } from "../session/session.js";
 import type { Event } from "../session/event-log.js";
-import type { RolloutItem } from "../session/rollout-item.js";
-import { reconstructFromRollout } from "../session/rollout-reconstruction.js";
 import type { TurnContext } from "../session/turn-context.js";
-import type {
-  SessionEditorInteraction,
-  SessionSubmitOptions,
-} from "../session/autonomous-mode.js";
-import { editorInteractionSystemPrompt } from "../session/editor-interaction.js";
-import type { CodePredictionSource } from "../services/code-prediction/types.js";
 import {
-  respondToSessionElicitation,
-  type SessionElicitationResponseParams,
-} from "../elicitation/respond.js";
+  editorInteractionSystemPrompt,
+} from "../session/editor-interaction.js";
 import type {
-  AgenCDaemonSessionNotification,
-  AgentRunStatus,
+  CodePredictionSource,
+} from "../services/code-prediction/types.js";
+import { respondToSessionElicitation } from "../elicitation/respond.js";
+import type {
   AgentStatus as DaemonAgentStatus,
   JsonObject,
-  JsonValue,
-  MessageContent,
   PermissionListResult,
-  SessionMcpServerConfig,
-  SessionPartialCompactFromMessageParams,
   SessionPartialCompactFromMessageResult,
   SessionRollbackCompactionResult,
   SessionExtendCompactionRollbackRetentionResult,
@@ -178,12 +125,7 @@ import type {
   SessionSnapshotResult,
   SessionTranscriptResult,
   SessionTranscriptV2Result,
-  SessionTranscriptV2TurnResult,
-  SessionHookConfigShape,
-  SessionHookValidationIssueShape,
-  SessionHookRunDiagnosticShape,
   SessionPermissionRuleMutationParams,
-  SessionPermissionRuleMutationResult,
   SessionShellExecuteParams,
   SessionShellExecuteResult,
 } from "./protocol/index.js";
@@ -191,906 +133,242 @@ import type { AgenCRealtimeThreadBinding } from "./realtime.js";
 import type { AgenCRealtimeCallClient } from "./realtime-transport.js";
 import type {
   RealtimeTransportConnection,
-  RealtimeTransportRequest,
 } from "../conversation/realtime/conversation.js";
-import type { RealtimeStartupContextSessionLike } from "../conversation/realtime/context.js";
-import {
-  JSON_RPC_VERSION,
-  MAX_SESSION_SHELL_RESULT_TEXT_UTF8_BYTES,
-} from "./protocol/index.js";
+import type {
+  RealtimeStartupContextSessionLike,
+} from "../conversation/realtime/context.js";
 import {
   createAgenCDaemonRuntimeAuthBackend,
   type AgenCDaemonRuntimeAuthBackend,
 } from "./provider-key-vending.js";
-import { isRecord } from "../utils/record.js";
-import type { ExecutionAdmissionKernel } from "../budget/execution-admission-kernel.js";
-import type { CsvAgentJobsRepositoryProvider } from "./csv-agent-jobs-authority.js";
+import type {
+  ExecutionAdmissionKernel,
+} from "../budget/execution-admission-kernel.js";
+import type {
+  CsvAgentJobsRepositoryProvider,
+} from "./csv-agent-jobs-authority.js";
 import {
-  EVENT_GAP_EVENT,
   RUN_RUNTIME_MODEL_VERBOSITIES,
-  RUN_RUNTIME_PERMISSION_MODES,
   RUN_RUNTIME_REASONING_EFFORTS,
   RUN_RUNTIME_SERVICE_TIERS,
-  RUN_RUNTIME_SETTINGS_CHANGE_REASONS,
-  type RunResumeReason,
   type RunRuntimeSettingsChangeReason,
   type RunRuntimeSettingsSnapshot,
-  type RunTerminalResult,
 } from "../contracts/run-contracts.js";
-import { cloneFrozenRuntimeSettingsSnapshot } from "../state/runtime-settings-snapshot.js";
-import type { ResumeRolloutDescriptorLease } from "../session/session-store.js";
-import type { AgentRuntimeOptions } from "../session/runtime-options.js";
-import { runWithAgentRuntimeOptions } from "../session/runtime-options.js";
 import {
-  applySessionExecutionAuthority,
-  executionAuthorityForPermissionContext,
-  sandboxExecutionBrokerAuthorityFromSessionAuthority,
-} from "../session/configuration.js";
-import { SandboxExecutionBroker } from "../sandbox/execution-broker.js";
-import { transitionSandboxExecutionBrokerAuthority } from "../sandbox/execution-lifecycle.js";
+  cloneFrozenRuntimeSettingsSnapshot,
+} from "../state/runtime-settings-snapshot.js";
+import { runWithAgentRuntimeOptions } from "../session/runtime-options.js";
 
-export interface AgenCBackgroundAgentStartParams {
-  readonly objective: string;
-  readonly cwd?: string;
-  readonly model?: string;
-  readonly provider?: string;
-  readonly profile?: string;
-  readonly configPath?: string;
-  readonly addDirs?: readonly string[];
-  readonly initialContent?: MessageContent;
-  readonly deferInitialTurn?: boolean;
-  readonly initialDisplayUserMessage?: string | null;
-  readonly initialEditorInteraction?: SessionEditorInteraction;
-  readonly metadata?: JsonObject;
-  readonly unattendedAllow: readonly string[];
-  readonly unattendedDeny: readonly string[];
-  readonly permissionMode?:
-    | "default"
-    | "plan"
-    | "acceptEdits"
-    | "bypassPermissions"
-    | "dontAsk"
-    | "auto";
-  readonly runtimeOptions: AgentRuntimeOptions;
-  /**
-   * Per-invocation env overrides forwarded from the CLI. Merged on
-   * top of `this.#env` so the user's latest `OPENAI_BASE_URL` /
-   * proxy / API key for THIS agent invocation wins over the frozen
-   * env snapshot captured when the daemon was launched.
-   */
-  readonly envOverrides?: { readonly [key: string]: string };
-}
-
-export interface AgenCBackgroundAgentStartResult {
-  readonly agentId: string;
-  readonly agentPath?: string;
-  /** Internal pre-publication rollback token; never serialized to clients. */
-  readonly restoreAttemptId?: string;
-  readonly startedAt: string;
-  readonly status: "running";
-  readonly rolloutPath?: string;
-  readonly rolloutDev?: string;
-  readonly rolloutIno?: string;
-}
-
-export interface AgenCBackgroundAgentRestoreParams {
-  readonly agentId: string;
-  readonly objective: string;
-  readonly cwd?: string;
-  readonly model?: string;
-  readonly provider?: string;
-  readonly profile?: string;
-  readonly configPath?: string;
-  readonly addDirs?: readonly string[];
-  readonly permissionMode?:
-    | "default"
-    | "plan"
-    | "acceptEdits"
-    | "bypassPermissions"
-    | "dontAsk"
-    | "auto";
-  readonly runtimeOptions: AgentRuntimeOptions;
-  /** Exact canonical rollout selected by the trusted CLI resolver. */
-  readonly resumeRolloutPath?: string;
-  /** One-shot daemon descriptor authority transferred into SessionStore. */
-  readonly resumeRolloutLease?: ResumeRolloutDescriptorLease;
-  /** Server-observed workspace identity, reproved inside bootstrap. */
-  readonly resumeCwdIdentity?: { readonly dev: string; readonly ino: string };
-  /** Daemon-held directory descriptor pinned across bootstrap acquisition. */
-  readonly resumeCwdFd?: number;
-  readonly envOverrides?: { readonly [key: string]: string };
-  /** User-requested continuation of a cleanly terminal canonical run. */
-  readonly reopenTerminalRun?: boolean;
-  /** Resume a daemon-suspended canonical run without changing its epoch. */
-  readonly resumeSuspendedRun?: boolean;
-  /** Internal suspension disposition; never inferred from caller prose. */
-  readonly suspendedResumeReason?: RunResumeReason;
-  /** Canonical open run still awaiting its durable first-input activation. */
-  readonly resumeStartupActivationPending?: boolean;
-  /** Complete canonical overlay restored before this generation accepts ingress. */
-  readonly runtimeSettings?: RunRuntimeSettingsSnapshot;
-  /** Exact-generation cleanup barrier for a user-selected disk continuation. */
-  readonly explicitColdResume?: boolean;
-  /** Opaque caller-owned generation token for pre-publication rollback only. */
-  readonly restoreAttemptId?: string;
-  readonly startedAt?: string;
-  readonly currentSessionId?: string;
-  readonly initialMessages?: ReadonlyArray<LLMMessage>;
-  readonly replayToolCalls?: readonly AgenCBackgroundAgentReplayToolCall[];
-  readonly onReplayToolResult?: (
-    result: AgenCBackgroundAgentReplayToolResult,
-  ) => void | Promise<void>;
-  readonly metadata?: JsonObject;
-}
-
-export interface AgenCBackgroundAgentReplayToolCall {
-  readonly callId: string;
-  readonly toolName: string;
-  readonly args: JsonValue;
-}
-
-export interface AgenCBackgroundAgentReplayToolResult {
-  readonly sessionId: string;
-  readonly callId: string;
-  readonly toolName: string;
-  readonly result: string;
-  readonly isError: boolean;
-  readonly terminalStatus?: "completed" | "failed" | "poisoned";
-  readonly recoveryCategory?: ToolRecoveryCategory;
-}
-
-export interface AgenCBackgroundAgentSnapshot {
-  readonly status: DaemonAgentStatus;
-  readonly lastActiveAt: string;
-  readonly metadata?: JsonObject;
-  /** Live daemon-owned session authority, captured after its durable commit. */
-  readonly runtimeSettings?: RunRuntimeSettingsSnapshot;
-  /** Canonical cursor for strict successor reconciliation on attached clients. */
-  readonly runtimeSettingsEventId?: string;
-  /** Present only after the canonical run_terminal event was fsync-committed. */
-  readonly terminal?: AgenCBackgroundAgentTerminalSnapshot;
-  /** Present only after the canonical run_suspended event was fsync-committed. */
-  readonly suspension?: AgenCBackgroundAgentSuspensionSnapshot;
-}
-
-export interface AgenCBackgroundAgentTerminalSnapshot {
-  readonly openedAt: string;
-  readonly epoch: number;
-  readonly eventId: string;
-  readonly rolloutPath: string;
-  readonly result: RunTerminalResult;
-}
-
-export interface AgenCBackgroundAgentSuspensionSnapshot {
-  readonly openedAt: string;
-  readonly epoch: number;
-  readonly eventId: string;
-  readonly sequence: number;
-  readonly rolloutPath: string;
-  readonly reason: "daemon_shutdown_idle";
-  readonly suspendedAt: string;
-}
-
-export type AgenCBackgroundAgentDaemonShutdownResult =
-  | {
-      readonly disposition: "suspended";
-      readonly suspension: AgenCBackgroundAgentSuspensionSnapshot;
-    }
-  | {
-      readonly disposition: "cancelled";
-      readonly terminal?: AgenCBackgroundAgentTerminalSnapshot;
-    };
-
-/**
- * The canonical suspension committed, but local daemon runtime teardown did
- * not complete cleanly. Callers must preserve the suspended projection while
- * still treating daemon cleanup as failed.
- */
-export class AgenCBackgroundAgentSuspensionShutdownError extends Error {
-  override readonly name = "AgenCBackgroundAgentSuspensionShutdownError";
-
-  constructor(
-    readonly suspension: AgenCBackgroundAgentSuspensionSnapshot,
-    cause: unknown,
-  ) {
-    super(
-      `run suspension ${suspension.eventId} committed but daemon runtime shutdown failed`,
-      { cause },
-    );
-  }
-}
-
-export interface AgenCBackgroundAgentCancellationPreparation {
-  readonly affectedRunIds: readonly string[];
-  readonly voidedHolds: number;
-  readonly heldUnknownHolds: number;
-}
-
-export interface AgenCBackgroundAgentTurnCancellationResult {
-  readonly cancelled: boolean;
-  readonly activeTurnId?: string;
-  readonly stale?: boolean;
-}
-
-export interface AgenCBackgroundAgentSessionEventBinding {
-  readonly sessionId: string;
-  readonly emit: (event: JsonObject) => void | Promise<void>;
-}
-
-export interface AgenCBackgroundAgentMessageParams {
-  readonly sessionId: string;
-  readonly content: MessageContent;
-  readonly originalContent: MessageContent;
-  readonly displayUserMessage?: string | null;
-  readonly editorInteraction?: SessionEditorInteraction;
-  readonly messageId: string;
-  readonly streamId: string;
-  readonly acceptedAt: string;
-  readonly ifBusy?: "reject";
-}
-
-export interface AgenCBackgroundAgentMessageResult {
-  readonly disposition: "started" | "duplicate";
-  readonly acceptedAt: string;
-  readonly duplicateState?: "completed" | "incomplete";
-  readonly turnId?: string;
-  readonly terminal?: AgenCBackgroundAgentMessageTerminal;
-}
-
-export interface AgenCBackgroundAgentMessageTerminal extends JsonObject {
-  readonly code: 0 | 1 | 130;
-  readonly message?: string;
-}
-
-export type AgenCBackgroundAgentMessageErrorCode =
-  "TURN_IN_PROGRESS" | "CLIENT_MESSAGE_ID_CONFLICT" | "PROMPT_BLOCKED";
-
-export class AgenCBackgroundAgentMessageError extends Error {
-  readonly code: AgenCBackgroundAgentMessageErrorCode;
-
-  constructor(code: AgenCBackgroundAgentMessageErrorCode, message: string) {
-    super(message);
-    this.name = "AgenCBackgroundAgentMessageError";
-    this.code = code;
-  }
-}
-
-const DAEMON_USER_PROMPT_PREPARED: unique symbol = Symbol(
-  "agenc.daemon-user-prompt-prepared",
-);
-
-type DaemonSessionSubmitOptions = SessionSubmitOptions & {
-  readonly [DAEMON_USER_PROMPT_PREPARED]?: true;
-};
-
-export interface AgenCBackgroundAgentClearSessionParams {
-  readonly sessionId: string;
-  readonly clearedAt: string;
-}
-
-export interface AgenCBackgroundAgentSnapshotSessionParams {
-  readonly sessionId: string;
-}
-
-export interface AgenCBackgroundAgentMcpAddServerParams {
-  readonly sessionId: string;
-  readonly config: SessionMcpServerConfig;
-}
-
-export interface AgenCBackgroundAgentMcpServerByNameParams {
-  readonly sessionId: string;
-  readonly serverName: string;
-}
-
-export interface AgenCBackgroundAgentPartialCompactParams {
-  readonly sessionId: string;
-  readonly messageOrdinal: number;
-  readonly direction: SessionPartialCompactFromMessageParams["direction"];
-  readonly feedback?: string;
-  readonly signal?: AbortSignal;
-}
-
-export interface AgenCBackgroundAgentRollbackCompactionParams {
-  readonly sessionId: string;
-  readonly attemptId: string;
-  readonly reviewedBranchTargetSessionId?: string;
-}
-
-export interface AgenCBackgroundAgentExtendCompactionRetentionParams {
-  readonly sessionId: string;
-  readonly attemptId: string;
-  readonly extendedUntilMs: number;
-}
-
-export interface AgenCBackgroundAgentConversationRewindParams {
-  readonly sessionId: string;
-  readonly messageOrdinal: number;
-  /** When aborted (request.cancel), refuse mid-flight rewind work (todo-108). */
-  readonly signal?: AbortSignal;
-}
-
-export interface AgenCBackgroundAgentSetModelParams {
-  readonly sessionId: string;
-  readonly model?: string;
-  readonly provider?: string;
-}
-
-export interface AgenCBackgroundAgentSetModelResult {
-  readonly applied: boolean;
-  readonly provider: string;
-  readonly model: string;
-  readonly runtimeSettingsEventId: string;
-  readonly summary: string;
-}
-
-export interface AgenCBackgroundAgentSetPermissionModeParams {
-  readonly sessionId: string;
-  readonly mode: string;
-  /** Bound only by an explicit allow-all decision in the tool approval flow. */
-  readonly bypassAuthority?: "operator_tool_approval";
-}
-
-export interface AgenCBackgroundAgentSetPermissionModeResult {
-  readonly applied: boolean;
-  readonly previousMode: string;
-  readonly mode: string;
-  /** Internal transaction hook; never serialized by session.setPermissionMode. */
-  readonly rollback?: () => Promise<void>;
-}
-
-export interface AgenCBackgroundAgentPermissionRuleMutationResult {
-  readonly applied: boolean;
-  readonly operation: SessionPermissionRuleMutationParams["operation"];
-  readonly behavior: SessionPermissionRuleMutationParams["behavior"];
-  readonly rule: string;
-  readonly sessionRules: SessionPermissionRuleMutationResult["sessionRules"];
-}
-
-export interface AgenCBackgroundAgentHooksStatusResult {
-  readonly available: boolean;
-  readonly sourcePath: string;
-  readonly disabled: boolean;
-  readonly hardSuppressed: boolean;
-  readonly effectiveDisabled: boolean;
-  readonly suppressionReason: "bare_mode" | "session_disabled" | null;
-  readonly issues: readonly SessionHookValidationIssueShape[];
-  readonly hooks: readonly SessionHookConfigShape[];
-  readonly diagnostics: readonly SessionHookRunDiagnosticShape[];
-}
-
-export interface AgenCBackgroundAgentSetHooksDisabledParams {
-  readonly disabled: boolean;
-}
-
-export interface AgenCBackgroundAgentSetHooksDisabledResult {
-  readonly applied: boolean;
-  readonly disabled: boolean;
-  readonly hardSuppressed: boolean;
-  readonly effectiveDisabled: boolean;
-  readonly suppressionReason: "bare_mode" | "session_disabled" | null;
-}
-
-function configuredHookExecutionState(runtime: {
-  isDisabled(): boolean;
-  isHardSuppressed(): boolean;
-  isExecutionSuppressed(): boolean;
-}): Omit<AgenCBackgroundAgentSetHooksDisabledResult, "applied"> {
-  const disabled = runtime.isDisabled();
-  const hardSuppressed = runtime.isHardSuppressed();
-  return {
-    disabled,
-    hardSuppressed,
-    effectiveDisabled: runtime.isExecutionSuppressed(),
-    suppressionReason: hardSuppressed
-      ? "bare_mode"
-      : disabled
-        ? "session_disabled"
-        : null,
-  };
-}
-
-export interface AgenCBackgroundAgentApplyConfigParams {
-  readonly sessionId: string;
-  readonly profile?: string;
-  readonly reload?: boolean;
-}
-
-export interface AgenCBackgroundAgentApplyConfigResult {
-  readonly applied: boolean;
-  readonly provider?: string;
-  readonly model?: string;
-  readonly runtimeSettingsEventId?: string;
-  readonly summary: string;
-}
-
-export interface AgenCBackgroundAgentToolDecisionParams {
-  readonly requestId: string;
-  readonly decision: ReviewDecision;
-}
-
-export interface AgenCBackgroundAgentToolCancelParams {
-  readonly requestId: string;
-  readonly reason?: string;
-}
-
-export type AgenCBackgroundAgentElicitationResponseParams =
-  SessionElicitationResponseParams;
-
-export interface AgenCBackgroundAgentRunner {
-  startAgent(
-    params: AgenCBackgroundAgentStartParams,
-  ): Promise<AgenCBackgroundAgentStartResult>;
-  getAgentSnapshot?(
-    agentId: string,
-  ): Promise<AgenCBackgroundAgentSnapshot | null>;
-  restoreAgent?(
-    params: AgenCBackgroundAgentRestoreParams,
-  ): Promise<boolean> | boolean;
-  /**
-   * Revoke an unpublished restored generation without poisoning its canonical
-   * open epoch. The opaque attempt id prevents cleanup from touching a newer
-   * generation installed under the same run id.
-   */
-  rollbackRestoredAgent?(
-    agentId: string,
-    restoreAttemptId: string,
-  ): Promise<void>;
-  /**
-   * Gate ingress and fsync cancellation/admission evidence while the Session
-   * journal listener is still live. `stopAgent` then owns quiescence and the
-   * terminal-tail append; the lifecycle projects legacy state afterward.
-   */
-  prepareAgentCancellation?(
-    agentId: string,
-    reason: string,
-  ): Promise<AgenCBackgroundAgentCancellationPreparation>;
-  stopAgent?(agentId: string, reason?: string): Promise<void>;
-  /** Daemon-only shutdown disposition; caller prose cannot select suspension. */
-  suspendIdleAgentForDaemonShutdown?(
-    agentId: string,
-  ): Promise<AgenCBackgroundAgentDaemonShutdownResult>;
-  attachAgentSessionEvents?(
-    agentId: string,
-    binding: AgenCBackgroundAgentSessionEventBinding,
-  ): Promise<void> | void;
-  submitAgentMessage?(
-    agentId: string,
-    params: AgenCBackgroundAgentMessageParams,
-  ): Promise<AgenCBackgroundAgentMessageResult>;
-  executeAgentShell?(
-    agentId: string,
-    params: SessionShellExecuteParams,
-    signal?: AbortSignal,
-  ): Promise<SessionShellExecuteResult>;
-  /** Resolve the live route without exposing the primary provider to callers. */
-  resolveCodePredictionSource?(
-    agentId: string,
-  ): Promise<CodePredictionSource> | CodePredictionSource;
-  clearAgentSession?(
-    agentId: string,
-    params: AgenCBackgroundAgentClearSessionParams,
-  ): Promise<void>;
-  snapshotAgentSession?(
-    agentId: string,
-    params: AgenCBackgroundAgentSnapshotSessionParams,
-  ): Promise<SessionSnapshotResult>;
-  getAgentSessionTranscript?(
-    agentId: string,
-    params: { readonly sessionId: string },
-  ): Promise<SessionTranscriptResult>;
-  getAgentSessionTranscriptV2?(
-    agentId: string,
-    params: { readonly sessionId: string },
-  ): Promise<SessionTranscriptV2Result>;
-  resolveLiveEffectReview?(
-    agentId: string,
-    params: ResolveDurableEffectReviewOptions,
-  ): Promise<ResolveDurableEffectReviewResult>;
-  addMcpServer?(
-    agentId: string,
-    params: AgenCBackgroundAgentMcpAddServerParams,
-  ): Promise<McpServerMutationResult>;
-  reconnectMcpServer?(
-    agentId: string,
-    params: AgenCBackgroundAgentMcpServerByNameParams,
-  ): Promise<McpServerMutationResult>;
-  enableMcpServer?(
-    agentId: string,
-    params: AgenCBackgroundAgentMcpServerByNameParams,
-  ): Promise<McpServerMutationResult>;
-  disableMcpServer?(
-    agentId: string,
-    params: AgenCBackgroundAgentMcpServerByNameParams,
-  ): Promise<McpServerMutationResult>;
-  getMcpStatus?(agentId: string): Promise<McpSurfaceSnapshot>;
-  partialCompactFromMessage?(
-    agentId: string,
-    params: AgenCBackgroundAgentPartialCompactParams,
-  ): Promise<SessionPartialCompactFromMessageResult>;
-  rollbackCompaction?(
-    agentId: string,
-    params: AgenCBackgroundAgentRollbackCompactionParams,
-  ): Promise<SessionRollbackCompactionResult>;
-  extendCompactionRollbackRetention?(
-    agentId: string,
-    params: AgenCBackgroundAgentExtendCompactionRetentionParams,
-  ): Promise<SessionExtendCompactionRollbackRetentionResult>;
-  rewindConversationToMessage?(
-    agentId: string,
-    params: AgenCBackgroundAgentConversationRewindParams,
-  ): Promise<SessionRewindConversationToMessageResult>;
-  previewFileRewind?(
-    agentId: string,
-    params: AgenCBackgroundAgentConversationRewindParams,
-  ): Promise<SessionPreviewFileRewindResult>;
-  rewindFilesToMessage?(
-    agentId: string,
-    params: AgenCBackgroundAgentConversationRewindParams,
-  ): Promise<SessionRewindFilesToMessageResult>;
-  setAgentModel?(
-    agentId: string,
-    params: AgenCBackgroundAgentSetModelParams,
-  ): Promise<AgenCBackgroundAgentSetModelResult>;
-  setAgentPermissionMode?(
-    agentId: string,
-    params: AgenCBackgroundAgentSetPermissionModeParams,
-  ): Promise<AgenCBackgroundAgentSetPermissionModeResult>;
-  mutateAgentPermissionRule?(
-    agentId: string,
-    params: SessionPermissionRuleMutationParams,
-  ): Promise<AgenCBackgroundAgentPermissionRuleMutationResult>;
-  getAgentHooksStatus?(
-    agentId: string,
-  ): Promise<AgenCBackgroundAgentHooksStatusResult>;
-  setAgentHooksDisabled?(
-    agentId: string,
-    params: AgenCBackgroundAgentSetHooksDisabledParams,
-  ): Promise<AgenCBackgroundAgentSetHooksDisabledResult>;
-  applyAgentConfig?(
-    agentId: string,
-    params: AgenCBackgroundAgentApplyConfigParams,
-  ): Promise<AgenCBackgroundAgentApplyConfigResult>;
-  resolveToolDecision?(
-    agentId: string,
-    params: AgenCBackgroundAgentToolDecisionParams,
-  ): Promise<boolean>;
-  cancelTool?(
-    agentId: string,
-    params: AgenCBackgroundAgentToolCancelParams,
-  ): Promise<boolean>;
-  /**
-   * Interrupt the agent's currently-running turn (if any). Resolves to
-   * `true` when an active turn was found and the agent's
-   * AbortController was fired; `false` when the agent was idle.
-   * Implementation MUST cascade to descendants so subagent turns are
-   * also stopped — see {@link AgentControl.interrupt}.
-   */
-  interruptAgentTurn?(agentId: string, reason: string): Promise<boolean>;
-  interruptAgentTurnIfMatches?(
-    agentId: string,
-    reason: string,
-    expectedTurnId: string,
-  ): Promise<AgenCBackgroundAgentTurnCancellationResult>;
-  /**
-   * Register a callback invoked once per agent immediately before the
-   * runner removes that agent from its `#active` registry on terminal
-   * status. The callback receives the final per-agent snapshot so the
-   * lifecycle layer can record the terminal-state transition before the
-   * underlying snapshot becomes unobservable. Without this hook,
-   * `getAgentSnapshot` returns null after cleanup and the lifecycle's
-   * lazy poll never observes the transition — leaving stale `running`
-   * entries in `agent.list`.
-   */
-  setOnActiveAgentTerminated?(
-    callback: (
-      agentId: string,
-      snapshot: AgenCBackgroundAgentSnapshot,
-    ) => void | Promise<void>,
-  ): void;
-  respondToElicitation?(
-    agentId: string,
-    params: AgenCBackgroundAgentElicitationResponseParams,
-  ): Promise<boolean>;
-  listPermissions?(agentId: string): Promise<PermissionListResult | null>;
-  resolveRealtimeThread?(
-    threadId: string,
-  ):
-    | AgenCRealtimeThreadBinding
-    | null
-    | Promise<AgenCRealtimeThreadBinding | null>;
-}
-
-export type AgenCRunAgentFunction = typeof runAgent;
-export type AgenCBootstrapFunction = (
-  options: BootstrapLocalRuntimeSessionOptions,
-) => Promise<LocalRuntimeBootstrap>;
-export type AgenCEnsureAgentControlFunction = typeof ensureAgentControl;
-export type AgenCBackgroundRealtimeTransportConnector = (
-  request: RealtimeTransportRequest,
-) => Promise<RealtimeTransportConnection> | RealtimeTransportConnection;
-
-interface ActiveBackgroundAgent {
-  readonly bootstrap: LocalRuntimeBootstrap;
-  readonly control: AgentControl;
-  readonly thread: ManagedThread;
-  status: DaemonAgentStatus;
-  readonly startedAt: string;
-  /** Opaque generation proof retained only until publication succeeds/fails. */
-  readonly restoreAttemptId?: string;
-  /** Current canonical lifecycle epoch, recovered from run_reopened events. */
-  runEpoch: number;
-  /** True once daemon delivery is sourced from the Session EventLog. */
-  canonicalEventBridgeInstalled: boolean;
-  /** Cached immediately after the fsync-committed run_terminal append. */
-  terminal?: AgenCBackgroundAgentTerminalSnapshot;
-  /** Cached immediately after the fsync-committed run_suspended append. */
-  suspension?: AgenCBackgroundAgentSuspensionSnapshot;
-  /** Set only by the daemon-only clean shutdown path. */
-  pendingSuspension?: {
-    readonly eventId: string;
-    readonly reason: "daemon_shutdown_idle";
-    readonly suspendedAt: string;
-  };
-  /** Closes every runner ingress route before idle state is observed. */
-  ingressClosed?: boolean;
-  /** Result selected before shutdown quiescence; committed at journal close. */
-  pendingTerminal?: RunTerminalResult;
-  /** Fsync-committed operator intent that precedes admission cancellation. */
-  cancellationRequest?: {
-    readonly eventId: string;
-    readonly sequence: number;
-    readonly reason: string;
-    readonly requestedAt: string;
-  };
-  /** Resume event that must be durably activated before ordinary input. */
-  pendingStartupActivationResumeEventId?: string;
-  /** Captured once so append ambiguity retries preserve exact evidence. */
-  pendingStartupActivationActivatedAt?: string;
-  runtimeSettings?: RunRuntimeSettingsSnapshot;
-  runtimeSettingsEventId?: string;
-  runtimeSettingsMutationQueue: Promise<void>;
-  /** Preserves a close-boundary append failure through fail-soft teardown. */
-  terminalCommitError?: unknown;
-  /** True when a real Session owns the before-close terminal finalizer. */
-  durableTerminalFinalizerInstalled: boolean;
-  unsubscribeDurableTerminalFinalizer?: () => void;
-  terminationNotified?: boolean;
-  lastActiveAt: string;
-  unsubscribeStatus?: () => void;
-  uninstallApprovalBridge?: () => void;
-  uninstallRuntimeSettingsPreCommit?: () => void;
-  unsubscribeElicitationEvents?: () => void;
-  unsubscribePhaseEvents?: () => void;
-  unsubscribeMcpSurfaceInvalidations?: () => void;
-  sessionBinding?: AgenCBackgroundAgentSessionEventBinding;
-  bufferedEvents: BackgroundAgentDaemonEvent[];
-  activeToolCallIds: Set<string>;
-  historyEpoch: string;
-  messageSubmission?: ActiveMessageSubmission;
-  messageSubmissionQueue: Promise<void>;
-  /** Resolves only after this exact generation has relinquished #active. */
-  cleanupComplete: Promise<void>;
-  pendingMessageSubmissionCount: number;
-  readonly messageSubmissionsById: Map<string, ActiveMessageSubmission>;
-  pendingShellExecutionCount: number;
-  readonly shellExecutionsById: Map<string, ActiveShellExecution>;
-  /**
-   * True when no initial turn was submitted at spawn (deferInitialTurn
-   * spawns and restored agents). Their thread sits in pending_init until
-   * the first accepted message initializes it, so pending_init must not
-   * count as busy for `ifBusy: "reject"` — rejecting there deadlocks the
-   * session: the message that would initialize the thread is the message
-   * being refused.
-   */
-  /**
-   * Per-agent emission serialization chain. `#emitOrBufferEvent` awaits
-   * an async-locked broadcast, so two fire-and-forget emits from a
-   * single callback (e.g. status + budget-usage) can otherwise complete
-   * out of order. Each emission is chained on this promise so events for
-   * ONE agent are delivered in arrival order; cross-agent concurrency is
-   * preserved because every ActiveBackgroundAgent owns its own chain.
-   * Mirrors AgenCStdioTransport's #dispatchChain (transport/stdio.ts).
-   */
-  dispatchChain: Promise<void>;
-}
-
-interface ActiveMessageSubmission {
-  readonly clientMessageId: string;
-  readonly contentFingerprint: string;
-  readonly streamId: string;
-  readonly acceptedAt: string;
-  turnId?: string;
-  assistantMessageOrdinal: number;
-  activeAssistantMessageId?: string;
-  terminal?: AgenCBackgroundAgentMessageTerminal;
-  readonly promise: Promise<AgenCBackgroundAgentMessageResult>;
-  settled: boolean;
-}
-
-interface ActiveShellExecution {
-  readonly commandFingerprint: string;
-  readonly promise: Promise<SessionShellExecuteResult>;
-  settled: boolean;
-}
-
-interface BackgroundAgentDaemonEvent {
-  /** Existing session/subscription correlation envelope. */
-  readonly id: string;
-  /** Canonical run-journal identity, distinct from the reusable envelope id. */
-  readonly eventId?: string;
-  /** Canonical rollout sequence when this event originated in Session.EventLog. */
-  readonly sequence?: number;
-  readonly type: string;
-  readonly payload?: JsonObject;
-  readonly messageId?: string;
-  readonly streamId?: string;
-  readonly acceptedAt?: string;
-  readonly runId?: string;
-  readonly historyEpoch?: string;
-  readonly turnId?: string;
-  readonly clientMessageId?: string;
-  /** Keep operation-scoped events out of model/run status projection. */
-  readonly statusProjection?: "session_only";
-}
-
-interface AgentTerminalUsage {
-  readonly inputTokens: number;
-  readonly outputTokens: number;
-  readonly totalTokens: number;
-  readonly costUsd: number;
-}
-
-/**
- * Upper bound on daemon events buffered for a single agent while no
- * session binding is attached (and on the per-agent `#pendingEvents`
- * detach buffer). A detached or pre-attach agent that never gets an
- * `agent.attach` would otherwise accumulate events on the heap without
- * limit. When the cap is exceeded the oldest events are dropped (FIFO
- * eviction) so the newest events — the ones most useful when the TUI
- * finally attaches — are retained. Mirrors
- * MAX_RETAINED_NOTIFICATIONS (tasks/lifecycle.ts) and the
- * per-session caps in agent-cli.ts / client-multiplexer.ts.
- */
-const MAX_BUFFERED_AGENT_EVENTS = 1_000;
-
-const BACKGROUND_RUNNER_GAP_SOURCE = "background_runner_retention";
-
-/**
- * Drops the oldest events in-place until `events` is within
- * {@link MAX_BUFFERED_AGENT_EVENTS}. Returns the same array so callers
- * can push then bound, matching `bufferSessionEvent` in
- * client-multiplexer.ts.
- */
-function boundBufferedAgentEvents(
-  events: BackgroundAgentDaemonEvent[],
-  runId?: string,
-): BackgroundAgentDaemonEvent[] {
-  const previousMarkers = events.filter(isBackgroundRunnerGapEvent);
-  const realEvents = events.filter(
-    (event) => !isBackgroundRunnerGapEvent(event),
-  );
-  const retired =
-    realEvents.length > MAX_BUFFERED_AGENT_EVENTS
-      ? realEvents.splice(0, realEvents.length - MAX_BUFFERED_AGENT_EVENTS)
-      : [];
-  const previousRetiredCount = previousMarkers.reduce(
-    (total, marker) => total + positiveInteger(marker.payload?.retiredCount),
-    0,
-  );
-  const retiredCount = previousRetiredCount + retired.length;
-  if (retiredCount === 0) {
-    events.splice(0, events.length, ...realEvents);
-    return events;
-  }
-
-  const priorAfterSequence = previousMarkers
-    .map((marker) => nonNegativeSequence(marker.payload?.afterSequence))
-    .find((value) => value !== undefined);
-  const previousCoordinatesUnknown = previousMarkers.some(
-    (marker) => marker.payload?.coordinatesAvailable === false,
-  );
-  const retiredSequences = retired.map((event) =>
-    positiveSequence(event.sequence),
-  );
-  const firstRetiredSequence = retiredSequences[0];
-  const allNewRetiredEventsSequenced = retiredSequences.every(
-    (value) => value !== undefined,
-  );
-  const afterSequence =
-    !previousCoordinatesUnknown && priorAfterSequence !== undefined
-      ? priorAfterSequence
-      : !previousCoordinatesUnknown &&
-          retired.length > 0 &&
-          allNewRetiredEventsSequenced &&
-          firstRetiredSequence !== undefined
-        ? firstRetiredSequence - 1
-        : undefined;
-  const firstAvailableSequence = positiveSequence(realEvents[0]?.sequence);
-  const coordinatesAvailable =
-    afterSequence !== undefined &&
-    afterSequence >= 0 &&
-    firstAvailableSequence !== undefined &&
-    firstAvailableSequence > afterSequence;
-  const resolvedRunId = runId ?? gapRunId(previousMarkers);
-  const marker: BackgroundAgentDaemonEvent = {
-    id: `runner-gap:${resolvedRunId ?? "unknown"}`,
-    type: EVENT_GAP_EVENT,
-    payload: {
-      kind: EVENT_GAP_EVENT,
-      reason: "retention",
-      source: BACKGROUND_RUNNER_GAP_SOURCE,
-      retiredCount,
-      coordinatesAvailable,
-      ...(resolvedRunId !== undefined ? { runId: resolvedRunId } : {}),
-      ...(coordinatesAvailable
-        ? { afterSequence, firstAvailableSequence }
-        : {}),
-    },
-  };
-  events.splice(0, events.length, marker, ...realEvents);
-  return events;
-}
-
-function isBackgroundRunnerGapEvent(
-  event: BackgroundAgentDaemonEvent,
-): boolean {
-  return (
-    event.type === EVENT_GAP_EVENT &&
-    event.payload?.source === BACKGROUND_RUNNER_GAP_SOURCE
-  );
-}
-
-function positiveSequence(value: unknown): number | undefined {
-  return typeof value === "number" && Number.isSafeInteger(value) && value > 0
-    ? value
-    : undefined;
-}
-
-function nonNegativeSequence(value: unknown): number | undefined {
-  return typeof value === "number" && Number.isSafeInteger(value) && value >= 0
-    ? value
-    : undefined;
-}
-
-function positiveInteger(value: unknown): number {
-  return typeof value === "number" && Number.isSafeInteger(value) && value > 0
-    ? value
-    : 0;
-}
-
-function gapRunId(
-  markers: readonly BackgroundAgentDaemonEvent[],
-): string | undefined {
-  return markers
-    .map((marker) => marker.payload?.runId)
-    .find(
-      (value): value is string => typeof value === "string" && value.length > 0,
-    );
-}
-
-export interface AgenCDelegateBackgroundAgentRunnerOptions {
-  readonly bootstrap?: AgenCBootstrapFunction;
-  readonly ensureAgentControl?: AgenCEnsureAgentControlFunction;
-  readonly authBackend?: AuthBackend;
-  readonly executionAdmissionKernel?: ExecutionAdmissionKernel;
-  readonly csvAgentJobsRepositories?: CsvAgentJobsRepositoryProvider;
-  readonly env?: NodeJS.ProcessEnv;
-  readonly argv?: readonly string[];
-  readonly now?: () => string;
-  readonly realtimeCallClient?: AgenCRealtimeCallClient;
-  readonly realtimeConnectTransport?: AgenCBackgroundRealtimeTransportConnector;
-  readonly onActiveAgentTerminated?: (
-    agentId: string,
-    snapshot: AgenCBackgroundAgentSnapshot,
-  ) => void | Promise<void>;
-}
-
-export type AgenCDelegateBackgroundAgentRunnerRuntimeConfig = Pick<
+import {
+  AgenCBackgroundAgentSuspensionShutdownError,
+  AgenCBackgroundAgentMessageError,
+  DAEMON_USER_PROMPT_PREPARED,
+  positiveSequence,
+  finiteNumber,
+  messageContentFingerprint,
+  metadataStringList,
+  isJsonObject,
+  isToolRecoveryCategory,
+  hashStable,
+} from "./background-agent-runner/shared.js";
+import type {
+  AgenCBackgroundAgentStartParams,
+  AgenCBackgroundAgentStartResult,
+  AgenCBackgroundAgentRestoreParams,
+  AgenCBackgroundAgentReplayToolCall,
+  AgenCBackgroundAgentReplayToolResult,
+  AgenCBackgroundAgentSnapshot,
+  AgenCBackgroundAgentDaemonShutdownResult,
+  AgenCBackgroundAgentCancellationPreparation,
+  AgenCBackgroundAgentTurnCancellationResult,
+  AgenCBackgroundAgentSessionEventBinding,
+  AgenCBackgroundAgentMessageParams,
+  AgenCBackgroundAgentMessageResult,
+  DaemonSessionSubmitOptions,
+  AgenCBackgroundAgentClearSessionParams,
+  AgenCBackgroundAgentSnapshotSessionParams,
+  AgenCBackgroundAgentMcpAddServerParams,
+  AgenCBackgroundAgentMcpServerByNameParams,
+  AgenCBackgroundAgentPartialCompactParams,
+  AgenCBackgroundAgentRollbackCompactionParams,
+  AgenCBackgroundAgentExtendCompactionRetentionParams,
+  AgenCBackgroundAgentConversationRewindParams,
+  AgenCBackgroundAgentSetModelParams,
+  AgenCBackgroundAgentSetModelResult,
+  AgenCBackgroundAgentSetPermissionModeParams,
+  AgenCBackgroundAgentSetPermissionModeResult,
+  AgenCBackgroundAgentPermissionRuleMutationResult,
+  AgenCBackgroundAgentHooksStatusResult,
+  AgenCBackgroundAgentSetHooksDisabledParams,
+  AgenCBackgroundAgentSetHooksDisabledResult,
+  AgenCBackgroundAgentApplyConfigParams,
+  AgenCBackgroundAgentApplyConfigResult,
+  AgenCBackgroundAgentToolDecisionParams,
+  AgenCBackgroundAgentToolCancelParams,
+  AgenCBackgroundAgentElicitationResponseParams,
+  AgenCBackgroundAgentRunner,
+  AgenCBootstrapFunction,
+  AgenCEnsureAgentControlFunction,
+  AgenCBackgroundRealtimeTransportConnector,
+  ActiveBackgroundAgent,
+  ActiveMessageSubmission,
+  ActiveShellExecution,
+  BackgroundAgentDaemonEvent,
   AgenCDelegateBackgroundAgentRunnerOptions,
-  "realtimeCallClient" | "realtimeConnectTransport"
-> & {
-  readonly authBackend: AuthBackend | undefined;
-};
+  AgenCDelegateBackgroundAgentRunnerRuntimeConfig,
+} from "./background-agent-runner/shared.js";
+import {
+  boundBufferedAgentEvents,
+  terminalUsageForActiveAgent,
+  pruneShellExecutionCache,
+  pruneMessageSubmissionCache,
+} from "./background-agent-runner/snapshot-retention.js";
+import {
+  phaseEventToProgressEvent,
+  canonicalSessionEventFromRecoveredProgress,
+  interruptedToolResultContent,
+} from "./background-agent-runner/progress-events.js";
+import {
+  sessionUserMessageEventFromDaemonEvent,
+  correlateDaemonEvent,
+  projectTelemetryErrorAsSessionOnly,
+  scopeDirectShellDaemonEvent,
+  notificationFromDaemonEvent,
+  daemonStatusFromRunTerminal,
+  daemonEventFromUnboundSessionEvent,
+  mapThreadStatus,
+  eventFromThreadStatus,
+} from "./background-agent-runner/daemon-events.js";
+import {
+  historyEpochFromRollout,
+  findPersistedMessageSubmission,
+  sessionTranscriptV2FromRollout,
+  currentRunEpochFromRollout,
+} from "./background-agent-runner/journal-reconstruction.js";
+import {
+  isRunnableActiveAgent,
+  isInterruptibleActiveAgent,
+  hasRuntimeActiveTurn,
+  hasOpenAgentDescendants,
+  runtimeActiveTurnId,
+  isClearInFlight,
+  shellSubmissionMessageId,
+  shellEventKey,
+  throwIfShellRequestAborted,
+  normalizeSessionShellResult,
+  clientMessageIdConflict,
+  messageContentToLlmParts,
+  commitDurableRunStartupActivation,
+  awaitTerminalStatus,
+  commitDurableRunCancellationRequest,
+  commitDurableRunTerminal,
+  commitDurableRunSuspension,
+  cancelledTerminalResult,
+  terminalResultFromThread,
+  messageContentToAgentInput,
+  submitStructuredAgentInput,
+  messageContentDisplayText,
+} from "./background-agent-runner/turn-lifecycle.js";
+import {
+  replayRecoveredToolCalls,
+  hydrateRecoveredSessionHistory,
+  resolvePermissionDecisionTimeoutMs,
+  readApprovalAgentId,
+} from "./background-agent-runner/tool-recovery.js";
+import {
+  configuredHookExecutionState,
+  runtimeWorkspaceRoot,
+  requireCanonicalRuntimeSettingsSupport,
+  failClosedDaemonRuntimeAuthority,
+  installDaemonPermissionAuthorityCoordinator,
+  prepareMcpAuthorityRefresh,
+  captureRuntimeSettings,
+  normalizeRuntimeSetting,
+  installRuntimeSettingsPreCommit,
+  withRuntimeSettingsMutation,
+  ensureInitialRuntimeSettings,
+  compensateRuntimeSettingsChange,
+  compensatePreparedRuntimeSettingsChange,
+  applyRestoredRuntimeSettings,
+  currentCanonicalRuntimeStateFromRollout,
+  commitDurableRuntimeSettingsChange,
+  prepareDurableRuntimeSettingsChange,
+  restoreBootstrapSelection,
+  runtimeSettingsWithRestoreOverrides,
+  buildBootstrapArgv,
+  installUnattendedPermissionPolicy,
+} from "./background-agent-runner/runtime-settings.js";
+import type {
+  PreparedRuntimeSettingsChange,
+} from "./background-agent-runner/runtime-settings.js";
+
+export {
+  AgenCBackgroundAgentSuspensionShutdownError,
+  AgenCBackgroundAgentMessageError,
+} from "./background-agent-runner/shared.js";
+export type {
+  AgenCBackgroundAgentStartParams,
+  AgenCBackgroundAgentStartResult,
+  AgenCBackgroundAgentRestoreParams,
+  AgenCBackgroundAgentReplayToolCall,
+  AgenCBackgroundAgentReplayToolResult,
+  AgenCBackgroundAgentSnapshot,
+  AgenCBackgroundAgentTerminalSnapshot,
+  AgenCBackgroundAgentSuspensionSnapshot,
+  AgenCBackgroundAgentDaemonShutdownResult,
+  AgenCBackgroundAgentCancellationPreparation,
+  AgenCBackgroundAgentTurnCancellationResult,
+  AgenCBackgroundAgentSessionEventBinding,
+  AgenCBackgroundAgentMessageParams,
+  AgenCBackgroundAgentMessageResult,
+  AgenCBackgroundAgentMessageTerminal,
+  AgenCBackgroundAgentMessageErrorCode,
+  AgenCBackgroundAgentClearSessionParams,
+  AgenCBackgroundAgentSnapshotSessionParams,
+  AgenCBackgroundAgentMcpAddServerParams,
+  AgenCBackgroundAgentMcpServerByNameParams,
+  AgenCBackgroundAgentPartialCompactParams,
+  AgenCBackgroundAgentRollbackCompactionParams,
+  AgenCBackgroundAgentExtendCompactionRetentionParams,
+  AgenCBackgroundAgentConversationRewindParams,
+  AgenCBackgroundAgentSetModelParams,
+  AgenCBackgroundAgentSetModelResult,
+  AgenCBackgroundAgentSetPermissionModeParams,
+  AgenCBackgroundAgentSetPermissionModeResult,
+  AgenCBackgroundAgentPermissionRuleMutationResult,
+  AgenCBackgroundAgentHooksStatusResult,
+  AgenCBackgroundAgentSetHooksDisabledParams,
+  AgenCBackgroundAgentSetHooksDisabledResult,
+  AgenCBackgroundAgentApplyConfigParams,
+  AgenCBackgroundAgentApplyConfigResult,
+  AgenCBackgroundAgentToolDecisionParams,
+  AgenCBackgroundAgentToolCancelParams,
+  AgenCBackgroundAgentElicitationResponseParams,
+  AgenCBackgroundAgentRunner,
+  AgenCRunAgentFunction,
+  AgenCBootstrapFunction,
+  AgenCEnsureAgentControlFunction,
+  AgenCBackgroundRealtimeTransportConnector,
+  AgenCDelegateBackgroundAgentRunnerOptions,
+  AgenCDelegateBackgroundAgentRunnerRuntimeConfig,
+} from "./background-agent-runner/shared.js";
+export {
+  managedTokenUsage,
+} from "./background-agent-runner/snapshot-retention.js";
+export {
+  phaseEventToProgressEvent,
+} from "./background-agent-runner/progress-events.js";
+export {
+  notificationFromDaemonEvent,
+  daemonEventFromUnboundSessionEvent,
+} from "./background-agent-runner/daemon-events.js";
+export {
+  sessionTranscriptV2FromRollout,
+} from "./background-agent-runner/journal-reconstruction.js";
+export {
+  resolvePermissionDecisionTimeoutMs,
+  planApprovalPayloadFields,
+} from "./background-agent-runner/tool-recovery.js";
 
 export class AgenCDelegateBackgroundAgentRunner implements AgenCBackgroundAgentRunner {
   readonly #bootstrap: AgenCBootstrapFunction;
@@ -5358,4270 +4636,8 @@ export class AgenCDelegateBackgroundAgentRunner implements AgenCBackgroundAgentR
   }
 }
 
-function sessionUserMessageEventFromDaemonEvent(
-  event: BackgroundAgentDaemonEvent,
-): Event | null {
-  if (
-    event.type !== "user_message" ||
-    event.payload === undefined ||
-    event.payload.message === undefined
-  ) {
-    return null;
-  }
-  return {
-    id: event.id,
-    msg: {
-      type: "user_message",
-      payload: {
-        message: event.payload.message as string | readonly LLMContentPart[],
-        ...(typeof event.payload.displayText === "string"
-          ? { displayText: event.payload.displayText }
-          : {}),
-        ...(Array.isArray(event.payload.images)
-          ? { images: stringArray(event.payload.images) }
-          : {}),
-        ...(typeof event.payload.queuedCommandUuid === "string"
-          ? { queuedCommandUuid: event.payload.queuedCommandUuid }
-          : {}),
-        ...(typeof event.messageId === "string"
-          ? { messageId: event.messageId }
-          : {}),
-        ...(typeof event.streamId === "string"
-          ? { streamId: event.streamId }
-          : {}),
-        ...(typeof event.acceptedAt === "string"
-          ? { acceptedAt: event.acceptedAt }
-          : {}),
-      },
-    },
-  };
-}
-
-function terminalUsageForActiveAgent(
-  active: ActiveBackgroundAgent,
-): AgentTerminalUsage {
-  const live = managedTokenUsage(active.thread);
-  return {
-    inputTokens: finiteNumber(live.inputTokens),
-    outputTokens: finiteNumber(live.outputTokens),
-    totalTokens: finiteNumber(live.totalTokens),
-    costUsd: agentCostUsd(active),
-  };
-}
-
-function agentCostUsd(active: ActiveBackgroundAgent): number {
-  const tokenUsage = managedTokenUsage(active.thread);
-  const model = activeAgentModel(active);
-  const provider = activeAgentProvider(active);
-  // LiveAgent currently exposes aggregate input/output token counters.
-  // Preserve that limited basis in the terminal usage snapshot without
-  // pretending cached/reasoning/search dimensions were observed.
-  const usage: ModelUsage = {
-    model,
-    ...(provider !== undefined ? { provider } : {}),
-    inputTokens: finiteNumber(tokenUsage.inputTokens),
-    outputTokens: finiteNumber(tokenUsage.outputTokens),
-    cachedInputTokens: 0,
-    cacheCreationInputTokens: 0,
-    reasoningOutputTokens: 0,
-    webSearchRequests: 0,
-    totalTokens: finiteNumber(tokenUsage.totalTokens),
-    turns: 0,
-  };
-  return computeUsdCost(usage, DEFAULT_MODEL_COSTS);
-}
-
-function activeAgentModel(active: ActiveBackgroundAgent): string {
-  return (
-    stringRecordField(active.thread.configSnapshot?.(), "model") ?? "agenc"
-  );
-}
-
-function activeAgentProvider(
-  active: ActiveBackgroundAgent,
-): string | undefined {
-  return (
-    stringRecordField(active.thread.configSnapshot?.(), "provider") ??
-    stringRecordField(active.thread.configSnapshot?.(), "model_provider")
-  );
-}
-
-function stringRecordField(
-  record: Record<string, unknown> | undefined,
-  key: string,
-): string | undefined {
-  const value = record?.[key];
-  return typeof value === "string" && value.trim().length > 0
-    ? value.trim()
-    : undefined;
-}
-
-function finiteNumber(value: unknown): number {
-  return typeof value === "number" && Number.isFinite(value) ? value : 0;
-}
-
-function isRunnableActiveAgent(active: ActiveBackgroundAgent): boolean {
-  return (
-    active.ingressClosed !== true &&
-    active.pendingTerminal === undefined &&
-    active.pendingSuspension === undefined
-  );
-}
-
-function isInterruptibleActiveAgent(active: ActiveBackgroundAgent): boolean {
-  return (
-    active.ingressClosed !== true &&
-    (active.pendingTerminal === undefined ||
-      active.cancellationRequest !== undefined)
-  );
-}
-
-interface ActiveTurnPeek {
-  unsafePeek?: () => unknown;
-}
-
-function hasRuntimeActiveTurn(
-  session: LocalRuntimeBootstrap["session"],
-): boolean {
-  const activeTurn = (session as unknown as { activeTurn?: ActiveTurnPeek })
-    .activeTurn;
-  return (
-    typeof activeTurn?.unsafePeek === "function" &&
-    activeTurn.unsafePeek() !== null
-  );
-}
-
-function hasOpenAgentDescendants(
-  control: AgentControl,
-  rootThreadId: string,
-): boolean {
-  const childrenByParent = control.liveThreadSpawnChildren();
-  const pending = [rootThreadId];
-  const visited = new Set<string>();
-  while (pending.length > 0) {
-    const parent = pending.pop()!;
-    if (visited.has(parent)) continue;
-    visited.add(parent);
-    const children = childrenByParent.get(parent) ?? [];
-    if (children.length > 0) return true;
-    for (const [childThreadId] of children) pending.push(childThreadId);
-  }
-  return false;
-}
-
-function runtimeActiveTurnId(
-  session: LocalRuntimeBootstrap["session"],
-): string | undefined {
-  const activeTurn = (session as unknown as { activeTurn?: ActiveTurnPeek })
-    .activeTurn;
-  if (typeof activeTurn?.unsafePeek !== "function") return undefined;
-  const value = activeTurn.unsafePeek();
-  if (!isJsonObject(value) || typeof value.turnId !== "string") {
-    return undefined;
-  }
-  return value.turnId;
-}
-
-function isClearInFlight(active: ActiveBackgroundAgent): boolean {
-  return (
-    active.pendingMessageSubmissionCount > 0 ||
-    active.pendingShellExecutionCount > 0 ||
-    active.messageSubmission !== undefined ||
-    hasRuntimeActiveTurn(active.bootstrap.session) ||
-    active.activeToolCallIds.size > 0
-  );
-}
-
-function correlateDaemonEvent(
-  active: ActiveBackgroundAgent,
-  event: BackgroundAgentDaemonEvent,
-): BackgroundAgentDaemonEvent {
-  const runId = active.thread.threadId;
-  if (event.type === "history_cleared" || event.type === "transcript_epoch") {
-    active.historyEpoch = historyEpochForBoundary(
-      runId,
-      event.eventId ?? event.id,
-    );
-    return { ...event, runId, historyEpoch: active.historyEpoch };
-  }
-
-  const submission = active.messageSubmission;
-  const isSubmissionUserMarker =
-    submission !== undefined &&
-    event.type === "user_message" &&
-    (event.messageId === submission.clientMessageId ||
-      event.payload?.messageId === submission.clientMessageId);
-  if (
-    submission !== undefined &&
-    submission.turnId === undefined &&
-    event.type !== "turn_started" &&
-    !isSubmissionUserMarker
-  ) {
-    // The durable user marker is visible before the queued input owns a
-    // runtime turn. Do not stamp tail events from the preceding turn with the
-    // new submission's identity while waiting for its turn_started boundary.
-    return {
-      ...event,
-      runId,
-      historyEpoch: active.historyEpoch,
-      ...(typeof event.payload?.turnId === "string"
-        ? { turnId: event.payload.turnId }
-        : {}),
-    };
-  }
-  if (
-    submission !== undefined &&
-    event.type === "turn_started" &&
-    typeof event.payload?.turnId === "string"
-  ) {
-    submission.turnId = event.payload.turnId;
-  }
-
-  const turnId =
-    submission?.turnId ??
-    (typeof event.payload?.turnId === "string"
-      ? event.payload.turnId
-      : undefined);
-  let messageId = event.messageId;
-  if (
-    submission !== undefined &&
-    (event.type === "agent_message_delta" || event.type === "agent_message")
-  ) {
-    if (submission.activeAssistantMessageId === undefined) {
-      submission.activeAssistantMessageId = assistantMessageId(
-        turnId ?? submission.clientMessageId,
-        submission.assistantMessageOrdinal,
-      );
-    }
-    messageId = submission.activeAssistantMessageId;
-    if (event.type === "agent_message") {
-      submission.activeAssistantMessageId = undefined;
-      submission.assistantMessageOrdinal += 1;
-    }
-  } else if (
-    submission === undefined &&
-    event.type === "agent_message" &&
-    messageId === undefined
-  ) {
-    messageId = `assistant:${event.eventId ?? event.id}`;
-  }
-  if (submission !== undefined) {
-    const terminal = messageTerminalFromDaemonEvent(event, submission.turnId);
-    if (terminal !== undefined) submission.terminal = terminal;
-  }
-
-  return {
-    ...event,
-    runId,
-    historyEpoch: active.historyEpoch,
-    ...(turnId !== undefined ? { turnId } : {}),
-    ...(submission !== undefined
-      ? { clientMessageId: submission.clientMessageId }
-      : {}),
-    ...(messageId !== undefined ? { messageId } : {}),
-  };
-}
-
-function messageTerminalFromDaemonEvent(
-  event: BackgroundAgentDaemonEvent,
-  expectedTurnId: string | undefined,
-): AgenCBackgroundAgentMessageTerminal | undefined {
-  const eventTurnId =
-    event.turnId ??
-    (typeof event.payload?.turnId === "string"
-      ? event.payload.turnId
-      : undefined);
-  if (
-    expectedTurnId !== undefined &&
-    eventTurnId !== undefined &&
-    eventTurnId !== expectedTurnId
-  ) {
-    return undefined;
-  }
-  if (event.type === "turn_complete") {
-    return {
-      code: 0,
-      ...(typeof event.payload?.lastAgentMessage === "string"
-        ? { message: event.payload.lastAgentMessage }
-        : {}),
-    };
-  }
-  if (event.type === "turn_aborted") {
-    return {
-      code: 130,
-      ...(typeof event.payload?.reason === "string"
-        ? { message: event.payload.reason }
-        : {}),
-    };
-  }
-  // `error` is session telemetry, not a turn closer. Stop-hook throws and
-  // similar sites emit it while the turn continues and later writes
-  // turn_complete / turn_aborted.
-  return undefined;
-}
-
-function assistantMessageId(turnId: string, ordinal: number): string {
-  return `assistant:${turnId}:${ordinal}`;
-}
-
-function historyEpochForBoundary(runId: string, boundaryId: string): string {
-  return `history:${runId}:${boundaryId}`;
-}
-
-function historyEpochFromRollout(
-  items: readonly RolloutItem[],
-  runId: string,
-): string {
-  return historyEpochForBoundary(
-    runId,
-    latestTranscriptBoundary(items)?.id ?? "initial",
-  );
-}
-
-interface TranscriptBoundary {
-  readonly index: number;
-  readonly id: string;
-  readonly kind: "cleared" | "replaced";
-  readonly sequence?: number;
-}
-
-function latestTranscriptBoundary(
-  items: readonly RolloutItem[],
-): TranscriptBoundary | undefined {
-  let latest: TranscriptBoundary | undefined;
-  for (const [index, item] of items.entries()) {
-    if (
-      item.type === "event_msg" &&
-      (item.payload.msg.type === "history_cleared" ||
-        item.payload.msg.type === "transcript_epoch")
-    ) {
-      latest = {
-        index,
-        id: canonicalEventId(item.payload),
-        kind:
-          item.payload.msg.type === "history_cleared" ? "cleared" : "replaced",
-        ...(positiveSequence(item.payload.seq) !== undefined
-          ? { sequence: positiveSequence(item.payload.seq) }
-          : {}),
-      };
-      continue;
-    }
-    // Compaction is deliberately NOT a transcript boundary.
-    //
-    // It changes what the MODEL sees; the person reading the transcript still
-    // sent every earlier message and expects to find them. Treating a commit
-    // as a boundary truncated the reloaded transcript to the compacted view
-    // and rendered the summary itself as a user message: live, a 13-turn
-    // session came back as two turns after an app relaunch, because that
-    // rollout contained two compaction commits and no explicit epoch event.
-    //
-    // A user-facing reset still truncates, and is still the only thing that
-    // does: `history_cleared` and `transcript_epoch` above. A partial compact
-    // or a rewind that means to reset the transcript emits `transcript_epoch`
-    // alongside its `compacted` item, so those keep working unchanged.
-  }
-  return latest;
-}
-
-function messageContentFingerprint(content: unknown): string {
-  return createHash("sha256")
-    .update(stableStringify(content) ?? "undefined", "utf8")
-    .digest("hex");
-}
-
-const MAX_RETAINED_SHELL_EXECUTIONS = 256;
-const SHELL_RESULT_TRUNCATION_MARKER = "\n[truncated]";
-
-function shellSubmissionMessageId(commandId: string): string {
-  return `shell:${commandId}`;
-}
-
-function shellEventKey(commandId: string): string {
-  return createHash("sha256")
-    .update(commandId, "utf8")
-    .digest("hex")
-    .slice(0, 32);
-}
-
-/**
- * Session `error` records are diagnostic events, not lifecycle boundaries.
- * Terminal run failures arrive through RunAgentProgressEvent.run_error and
- * update the lifecycle status there. Keep the diagnostic event visible
- * without letting it poison the keep-alive session first.
- */
-function projectTelemetryErrorAsSessionOnly(
-  event: BackgroundAgentDaemonEvent,
-): BackgroundAgentDaemonEvent {
-  if (event.type === "error") {
-    return { ...event, statusProjection: "session_only" };
-  }
-  return event;
-}
-
-function scopeDirectShellDaemonEvent(
-  active: ActiveBackgroundAgent,
-  event: BackgroundAgentDaemonEvent,
-): BackgroundAgentDaemonEvent {
-  const payload = event.payload;
-  const correlationIds = [
-    event.id,
-    payload?.callId,
-    payload?.queuedCommandUuid,
-  ];
-  if (
-    !correlationIds.some(
-      (candidate) =>
-        typeof candidate === "string" &&
-        active.shellExecutionsById.has(candidate),
-    )
-  ) {
-    return event;
-  }
-  return { ...event, statusProjection: "session_only" };
-}
-
-function throwIfShellRequestAborted(signal: AbortSignal | undefined): void {
-  if (signal?.aborted !== true) return;
-  if (signal.reason instanceof Error) throw signal.reason;
-  throw new Error(
-    typeof signal.reason === "string" && signal.reason.length > 0
-      ? signal.reason
-      : "Shell command cancelled",
-  );
-}
-
-function pruneShellExecutionCache(
-  cache: Map<string, ActiveShellExecution>,
-): void {
-  if (cache.size <= MAX_RETAINED_SHELL_EXECUTIONS) return;
-  for (const [commandId, execution] of cache) {
-    if (!execution.settled) continue;
-    cache.delete(commandId);
-    if (cache.size <= MAX_RETAINED_SHELL_EXECUTIONS) return;
-  }
-}
-
-function recordValue(value: unknown): Record<string, unknown> {
-  return value !== null && typeof value === "object" && !Array.isArray(value)
-    ? (value as Record<string, unknown>)
-    : {};
-}
-
-function boundShellResultText(value: string): {
-  readonly value: string;
-  readonly truncated: boolean;
-} {
-  const encoded = Buffer.from(value, "utf8");
-  if (encoded.byteLength <= MAX_SESSION_SHELL_RESULT_TEXT_UTF8_BYTES) {
-    return { value, truncated: false };
-  }
-  const marker = Buffer.from(SHELL_RESULT_TRUNCATION_MARKER, "utf8");
-  let end = Math.max(
-    0,
-    MAX_SESSION_SHELL_RESULT_TEXT_UTF8_BYTES - marker.byteLength,
-  );
-  while (end > 0 && (encoded[end] & 0xc0) === 0x80) end -= 1;
-  return {
-    value: `${encoded.subarray(0, end).toString("utf8")}${SHELL_RESULT_TRUNCATION_MARKER}`,
-    truncated: true,
-  };
-}
-
-function normalizeSessionShellResult(
-  commandId: string,
-  dispatch: ToolDispatchResult,
-): SessionShellExecuteResult {
-  const metadata = recordValue(dispatch.metadata);
-  const codeMode = recordValue(dispatch.codeModeResult);
-  const metadataStdout =
-    typeof metadata.stdout === "string" ? metadata.stdout : undefined;
-  const codeModeOutput =
-    typeof codeMode.output === "string" ? codeMode.output : undefined;
-  const rawStdout =
-    metadataStdout ??
-    codeModeOutput ??
-    (dispatch.isError === true ? "" : dispatch.content);
-  const rawStderr =
-    typeof metadata.stderr === "string"
-      ? metadata.stderr
-      : dispatch.isError === true && rawStdout.length === 0
-        ? dispatch.content
-        : "";
-  const metadataExitCode = metadata.exitCode;
-  const codeModeExitCode = codeMode.exit_code;
-  const exitCode =
-    typeof metadataExitCode === "number" &&
-    Number.isSafeInteger(metadataExitCode)
-      ? metadataExitCode
-      : typeof codeModeExitCode === "number" &&
-          Number.isSafeInteger(codeModeExitCode)
-        ? codeModeExitCode
-        : null;
-  const timedOut = metadata.timedOut === true || codeMode.timed_out === true;
-  const content = boundShellResultText(dispatch.content);
-  const stdout = boundShellResultText(rawStdout);
-  const stderr = boundShellResultText(rawStderr);
-  const truncated =
-    metadata.truncated === true ||
-    content.truncated ||
-    stdout.truncated ||
-    stderr.truncated;
-  return {
-    commandId,
-    content: content.value,
-    stdout: stdout.value,
-    stderr: stderr.value,
-    exitCode,
-    timedOut,
-    truncated,
-    isError:
-      dispatch.isError === true ||
-      timedOut ||
-      (exitCode !== null && exitCode !== 0),
-  };
-}
-
-interface PersistedMessageSubmission {
-  readonly contentFingerprint: string;
-  readonly acceptedAt?: string;
-  readonly turnId?: string;
-  readonly terminal?: AgenCBackgroundAgentMessageTerminal;
-}
-
-function findPersistedMessageSubmission(
-  items: readonly RolloutItem[],
-  clientMessageId: string,
-): PersistedMessageSubmission | undefined {
-  let match: PersistedMessageSubmission | undefined;
-  for (const item of items) {
-    if (item.type !== "event_msg") continue;
-    const event = item.payload;
-    if (
-      match === undefined &&
-      ((event.msg.type === "user_message" &&
-        event.msg.payload.messageId === clientMessageId) ||
-        (event.msg.type === "message_submission" &&
-          event.msg.payload.messageId === clientMessageId))
-    ) {
-      const contentFingerprint =
-        event.msg.type === "message_submission"
-          ? event.msg.payload.contentFingerprint
-          : messageContentFingerprint(event.msg.payload.message);
-      const acceptedAt = event.msg.payload.acceptedAt;
-      match = {
-        contentFingerprint,
-        ...(acceptedAt !== undefined ? { acceptedAt } : {}),
-      };
-      continue;
-    }
-    if (match === undefined) continue;
-    // A user_message starts the next admitted submission. Never let a
-    // crash-tail retry inherit that later submission's turn_started or
-    // terminal outcome.
-    if (
-      event.msg.type === "user_message" ||
-      event.msg.type === "message_submission"
-    ) {
-      return match;
-    }
-    if (event.msg.type === "turn_started" && match.turnId === undefined) {
-      match = { ...match, turnId: event.msg.payload.turnId };
-      continue;
-    }
-    if (match.turnId === undefined) continue;
-    const terminal = messageTerminalFromEvent(event.msg, match.turnId);
-    if (terminal !== undefined) {
-      return { ...match, terminal };
-    }
-  }
-  return match;
-}
-
-function messageTerminalFromEvent(
-  event: Event["msg"],
-  expectedTurnId: string | undefined,
-): AgenCBackgroundAgentMessageTerminal | undefined {
-  if (event.type === "turn_complete") {
-    if (
-      expectedTurnId !== undefined &&
-      event.payload.turnId !== expectedTurnId
-    ) {
-      return undefined;
-    }
-    return {
-      code: 0,
-      ...(event.payload.lastAgentMessage !== undefined
-        ? { message: event.payload.lastAgentMessage }
-        : {}),
-    };
-  }
-  if (event.type === "turn_aborted") {
-    if (
-      expectedTurnId !== undefined &&
-      event.payload.turnId !== undefined &&
-      event.payload.turnId !== expectedTurnId
-    ) {
-      return undefined;
-    }
-    return { code: 130, message: event.payload.reason };
-  }
-  // Same contract as the live bridge: only turn_complete / turn_aborted
-  // close a submission. A mid-turn `error` must not make an idempotent
-  // retry report completed-with-failure while turn_complete is still
-  // ahead in the journal.
-  return undefined;
-}
-
-function clientMessageIdConflict(
-  clientMessageId: string,
-): AgenCBackgroundAgentMessageError {
-  return new AgenCBackgroundAgentMessageError(
-    "CLIENT_MESSAGE_ID_CONFLICT",
-    `clientMessageId ${clientMessageId} was already used for different content`,
-  );
-}
-
-const MAX_MESSAGE_SUBMISSION_CACHE = 1_024;
-
-function pruneMessageSubmissionCache(
-  submissions: Map<string, ActiveMessageSubmission>,
-): void {
-  if (submissions.size <= MAX_MESSAGE_SUBMISSION_CACHE) return;
-  for (const [clientMessageId, submission] of submissions) {
-    if (!submission.settled) continue;
-    submissions.delete(clientMessageId);
-    if (submissions.size <= MAX_MESSAGE_SUBMISSION_CACHE) return;
-  }
-}
-
-function canonicalEventId(event: Event): string {
-  return (
-    event.eventId ??
-    (event.seq !== undefined
-      ? `legacy-event:${event.seq}:${event.id}`
-      : event.id)
-  );
-}
-
-interface MutableTranscriptV2Message extends JsonObject {
-  messageId: string;
-  commitEventId: string;
-  role: "user" | "assistant";
-  text: string;
-  turnId?: string;
-  clientMessageId?: string;
-  committedSequence: number;
-}
-
-/** Running totals for the turn currently open in the canonical scan. */
-interface OpenTurnAccumulator {
-  startedAt?: number;
-  sawUsage: boolean;
-  inputTokens: number;
-  outputTokens: number;
-  totalTokens: number;
-  model?: string;
-  provider?: string;
-}
-
-function nonNegativeFinite(value: unknown): number | undefined {
-  return typeof value === "number" && Number.isFinite(value) && value >= 0
-    ? value
-    : undefined;
-}
-
-/**
- * A turn's terminal row: timing from the terminal event itself (falling back
- * to the started/completed stamps for rollouts written before `durationMs`),
- * usage summed from the token_count events the turn enclosed.
- */
-function closedTurnResult(
-  turnId: string,
-  sequence: number,
-  msg: {
-    readonly type: string;
-    readonly payload: {
-      readonly turnId?: string;
-      readonly durationMs?: number;
-      readonly completedAt?: number;
-    };
-  },
-  open: OpenTurnAccumulator,
-): SessionTranscriptV2TurnResult {
-  const outcome = msg.type === "turn_aborted" ? "aborted" : "completed";
-  let durationMs: number | undefined;
-  if (msg.type === "turn_complete") {
-    durationMs = nonNegativeFinite(msg.payload.durationMs);
-    if (durationMs === undefined) {
-      const completedAt = nonNegativeFinite(msg.payload.completedAt);
-      if (completedAt !== undefined && open.startedAt !== undefined) {
-        durationMs = nonNegativeFinite(completedAt - open.startedAt);
-      }
-    }
-  }
-  return {
-    turnId,
-    committedSequence: sequence,
-    outcome,
-    ...(durationMs !== undefined ? { durationMs } : {}),
-    ...(open.sawUsage
-      ? {
-          inputTokens: open.inputTokens,
-          outputTokens: open.outputTokens,
-          totalTokens: open.totalTokens,
-        }
-      : {}),
-    ...(open.model !== undefined ? { model: open.model } : {}),
-    ...(open.provider !== undefined ? { provider: open.provider } : {}),
-  };
-}
-
-export function sessionTranscriptV2FromRollout(
-  items: readonly RolloutItem[],
-  sessionId: string,
-  runId: string,
-  activeTurn?: { readonly turnId: string; readonly clientMessageId: string },
-): SessionTranscriptV2Result {
-  const boundary = latestTranscriptBoundary(items);
-  const boundaryIndex = boundary?.index ?? -1;
-  const boundaryId = boundary?.id ?? "initial";
-  let asOfSequence = 0;
-  for (const item of items) {
-    if (item.type !== "event_msg") continue;
-    const event = item.payload;
-    if (
-      event.seq !== undefined &&
-      Number.isSafeInteger(event.seq) &&
-      event.seq > asOfSequence
-    ) {
-      asOfSequence = event.seq;
-    }
-  }
-
-  const messages: MutableTranscriptV2Message[] = [];
-  const turnResults: SessionTranscriptV2TurnResult[] = [];
-  let currentTurnId: string | undefined;
-  let currentClientMessageId: string | undefined;
-  let openTurn: OpenTurnAccumulator | undefined;
-  let pendingUserIndex: number | undefined;
-  let pendingClientMessageId: string | undefined;
-  const assistantOrdinals = new Map<string, number>();
-
-  if (boundary?.kind === "replaced") {
-    const replacement = reconstructFromRollout(
-      items.slice(0, boundary.index + 1),
-    ).history;
-    const replacementSequence =
-      boundary.sequence ?? maxEventSequence(items.slice(0, boundary.index + 1));
-    let ordinal = 0;
-    for (const item of replacement) {
-      if (item.role !== "user" && item.role !== "assistant") continue;
-      const text = responseItemDisplayText(item.content);
-      if (text.length === 0) continue;
-      const messageId = `replacement:${boundary.id}:${ordinal}`;
-      messages.push({
-        messageId,
-        commitEventId: messageId,
-        role: item.role,
-        text,
-        committedSequence: replacementSequence,
-      });
-      ordinal += 1;
-    }
-  }
-
-  const transcriptStartIndex = boundaryIndex + 1;
-  const firstCanonicalTranscriptIndex = items.findIndex(
-    (item, index) =>
-      index >= transcriptStartIndex &&
-      item.type === "event_msg" &&
-      (item.payload.msg.type === "user_message" ||
-        item.payload.msg.type === "message_submission" ||
-        item.payload.msg.type === "agent_message"),
-  );
-  const legacyEndIndex =
-    firstCanonicalTranscriptIndex < 0
-      ? items.length
-      : firstCanonicalTranscriptIndex;
-  let legacyOrdinal = 0;
-  for (let index = transcriptStartIndex; index < legacyEndIndex; index += 1) {
-    const item = items[index]!;
-    if (item.type !== "response_item") continue;
-    if (item.payload.role !== "user" && item.payload.role !== "assistant") {
-      continue;
-    }
-    const text = responseItemDisplayText(item.payload.content);
-    if (text.length === 0) continue;
-    const messageId = `legacy:${boundaryId}:${legacyOrdinal}`;
-    messages.push({
-      messageId,
-      commitEventId: messageId,
-      role: item.payload.role,
-      text,
-      committedSequence: 0,
-    });
-    legacyOrdinal += 1;
-  }
-
-  // Scan canonical turn context from the epoch boundary, not merely from the
-  // first visible transcript row. Hidden-user submissions deliberately have
-  // no user_message, so their preceding turn_started is the only durable
-  // source for the assistant message identity.
-  for (let index = transcriptStartIndex; index < items.length; index += 1) {
-    const item = items[index]!;
-    if (item.type !== "event_msg") continue;
-    const event = item.payload;
-    const sequence = positiveSequence(event.seq);
-    if (sequence === undefined) continue;
-    if (event.msg.type === "message_submission") {
-      pendingUserIndex = undefined;
-      pendingClientMessageId = event.msg.payload.messageId;
-      continue;
-    }
-    if (event.msg.type === "user_message") {
-      const text =
-        event.msg.payload.displayText ??
-        unknownMessageContentDisplayText(event.msg.payload.message);
-      if (text.length === 0) continue;
-      const commitEventId = canonicalEventId(event);
-      const clientMessageId = event.msg.payload.messageId;
-      messages.push({
-        messageId: clientMessageId ?? `message:${commitEventId}`,
-        commitEventId,
-        role: "user",
-        text,
-        ...(clientMessageId !== undefined ? { clientMessageId } : {}),
-        committedSequence: sequence,
-      });
-      pendingUserIndex = messages.length - 1;
-      pendingClientMessageId = clientMessageId;
-      continue;
-    }
-    if (event.msg.type === "turn_started") {
-      currentTurnId = event.msg.payload.turnId;
-      if (pendingUserIndex !== undefined) {
-        messages[pendingUserIndex]!.turnId = currentTurnId;
-        pendingUserIndex = undefined;
-      }
-      currentClientMessageId = pendingClientMessageId;
-      pendingClientMessageId = undefined;
-      // A dangling previous turn has no terminal event and therefore no
-      // result row; the fresh accumulator simply replaces it.
-      const startedAt = nonNegativeFinite(event.msg.payload.startedAt);
-      openTurn = {
-        sawUsage: false,
-        inputTokens: 0,
-        outputTokens: 0,
-        totalTokens: 0,
-        ...(startedAt !== undefined ? { startedAt } : {}),
-      };
-      continue;
-    }
-    if (event.msg.type === "token_count") {
-      if (currentTurnId === undefined || openTurn === undefined) continue;
-      const payload = event.msg.payload;
-      const input = nonNegativeFinite(payload.promptTokens);
-      const output = nonNegativeFinite(payload.completionTokens);
-      const total = nonNegativeFinite(payload.totalTokens);
-      if (input === undefined && output === undefined && total === undefined) {
-        continue;
-      }
-      openTurn.sawUsage = true;
-      openTurn.inputTokens += input ?? 0;
-      openTurn.outputTokens += output ?? 0;
-      openTurn.totalTokens += total ?? 0;
-      if (typeof payload.model === "string" && payload.model.length > 0) {
-        openTurn.model = payload.model;
-      }
-      if (typeof payload.provider === "string" && payload.provider.length > 0) {
-        openTurn.provider = payload.provider;
-      }
-      continue;
-    }
-    if (event.msg.type === "agent_message") {
-      const commitEventId = canonicalEventId(event);
-      const correlationId = currentTurnId ?? commitEventId;
-      const ordinal = assistantOrdinals.get(correlationId) ?? 0;
-      assistantOrdinals.set(correlationId, ordinal + 1);
-      const text = event.msg.payload.message;
-      // Empty durable commits are not transcript rows, but they still occupy
-      // an ordinal because the live identity allocator observes them.
-      if (text.length === 0) continue;
-      messages.push({
-        messageId:
-          currentTurnId === undefined
-            ? `assistant:${commitEventId}`
-            : assistantMessageId(currentTurnId, ordinal),
-        commitEventId,
-        role: "assistant",
-        text,
-        ...(currentTurnId !== undefined ? { turnId: currentTurnId } : {}),
-        ...(currentClientMessageId !== undefined
-          ? { clientMessageId: currentClientMessageId }
-          : {}),
-        committedSequence: sequence,
-      });
-      continue;
-    }
-    if (
-      event.msg.type === "turn_complete" ||
-      event.msg.type === "turn_aborted"
-    ) {
-      const terminalTurnId =
-        "turnId" in event.msg.payload &&
-        typeof event.msg.payload.turnId === "string"
-          ? event.msg.payload.turnId
-          : undefined;
-      if (
-        (currentTurnId === undefined && pendingUserIndex !== undefined) ||
-        (currentTurnId !== undefined &&
-          terminalTurnId !== undefined &&
-          terminalTurnId !== currentTurnId)
-      ) {
-        continue;
-      }
-      if (currentTurnId !== undefined && openTurn !== undefined) {
-        turnResults.push(
-          closedTurnResult(currentTurnId, sequence, event.msg, openTurn),
-        );
-      }
-      currentTurnId = undefined;
-      currentClientMessageId = undefined;
-      openTurn = undefined;
-    }
-  }
-
-  return {
-    schemaVersion: 2,
-    sessionId,
-    runId,
-    historyEpoch: historyEpochForBoundary(runId, boundaryId),
-    asOfSequence,
-    messages,
-    ...(activeTurn !== undefined ? { activeTurn } : {}),
-    ...(turnResults.length > 0 ? { turnResults } : {}),
-  };
-}
-
-function maxEventSequence(items: readonly RolloutItem[]): number {
-  let max = 0;
-  for (const item of items) {
-    if (item.type !== "event_msg") continue;
-    const sequence = positiveSequence(item.payload.seq);
-    if (sequence !== undefined) max = Math.max(max, sequence);
-  }
-  return max;
-}
-
-function responseItemDisplayText(content: unknown): string {
-  if (typeof content === "string") return content;
-  if (!Array.isArray(content)) return "";
-  return content
-    .map((part) =>
-      isJsonObject(part) && typeof part.text === "string" ? part.text : "",
-    )
-    .join("");
-}
-
-function unknownMessageContentDisplayText(content: unknown): string {
-  if (typeof content === "string") return content;
-  if (!Array.isArray(content)) return "";
-  return content
-    .map((part) => {
-      if (!isJsonObject(part)) return "";
-      if (typeof part.text === "string") return part.text;
-      if (part.type === "image") return "[image]";
-      if (part.type === "document") return "[document]";
-      return "";
-    })
-    .filter((part) => part.length > 0)
-    .join("\n");
-}
-
 async function unavailableRealtimeTransport(): Promise<RealtimeTransportConnection> {
   throw new Error("realtime transport connector is unavailable");
-}
-
-export function notificationFromDaemonEvent(
-  sessionId: string,
-  agentId: string,
-  event: BackgroundAgentDaemonEvent,
-): AgenCDaemonSessionNotification {
-  const base = eventBaseParams(sessionId, agentId, event);
-  const payload = event.payload;
-  if (
-    event.type === "mcp_status_changed" &&
-    isJsonObject(payload) &&
-    typeof payload.revision === "number" &&
-    Number.isSafeInteger(payload.revision) &&
-    payload.revision >= 0
-  ) {
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.mcp_status_changed",
-      params: {
-        sessionId,
-        revision: payload.revision,
-      },
-    };
-  }
-  if (
-    event.type === "agent_message_delta" &&
-    isJsonObject(payload) &&
-    typeof payload.delta === "string"
-  ) {
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.message_chunk",
-      params: {
-        ...base,
-        ...(event.messageId !== undefined
-          ? { messageId: event.messageId }
-          : {}),
-        ...(event.streamId !== undefined ? { streamId: event.streamId } : {}),
-        delta: payload.delta,
-      },
-    };
-  }
-  if (
-    event.type === "tool_call_started" &&
-    isJsonObject(payload) &&
-    typeof payload.callId === "string" &&
-    typeof payload.toolName === "string"
-  ) {
-    const input = toolRequestInputFromPayload(payload);
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.tool_request",
-      params: {
-        ...base,
-        requestId: payload.callId,
-        toolName: payload.toolName,
-        ...(input !== undefined ? { input } : {}),
-        ...(isToolRecoveryCategory(payload.recoveryCategory)
-          ? { recoveryCategory: payload.recoveryCategory }
-          : {}),
-      },
-    };
-  }
-  if (
-    event.type === "request_permissions" &&
-    isJsonObject(payload) &&
-    typeof payload.callId === "string"
-  ) {
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.permission_request",
-      params: {
-        ...base,
-        requestId: payload.callId,
-        ...(typeof payload.toolName === "string"
-          ? { toolName: payload.toolName }
-          : {}),
-        ...(typeof payload.turnId === "string"
-          ? { turnId: payload.turnId }
-          : {}),
-        permissions: stringArray(payload.permissions),
-        ...(payload.input !== undefined ? { input: payload.input } : {}),
-        ...(typeof payload.reason === "string"
-          ? { reason: payload.reason }
-          : {}),
-        ...(typeof payload.planContent === "string"
-          ? { planContent: payload.planContent }
-          : {}),
-        ...(typeof payload.planFilePath === "string"
-          ? { planFilePath: payload.planFilePath }
-          : {}),
-      },
-    };
-  }
-  if (
-    event.type === "request_user_input" &&
-    isJsonObject(payload) &&
-    typeof payload.callId === "string" &&
-    typeof payload.turnId === "string" &&
-    Array.isArray(payload.questions)
-  ) {
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.user_input_request",
-      params: {
-        ...base,
-        requestId:
-          typeof payload.requestId === "string"
-            ? payload.requestId
-            : payload.callId,
-        callId: payload.callId,
-        turnId: payload.turnId,
-        questions: jsonObjectArray(payload.questions),
-        ...(isJsonObject(payload.clientAction)
-          ? { clientAction: payload.clientAction }
-          : {}),
-      },
-    };
-  }
-  if (
-    event.type === "mcp_elicitation_request" &&
-    isJsonObject(payload) &&
-    typeof payload.serverName === "string" &&
-    (typeof payload.requestId === "string" ||
-      typeof payload.requestId === "number") &&
-    typeof payload.turnId === "string" &&
-    isJsonObject(payload.request)
-  ) {
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.mcp_elicitation_request",
-      params: {
-        ...base,
-        requestId: payload.requestId,
-        serverName: payload.serverName,
-        turnId: payload.turnId,
-        request: payload.request,
-      },
-    };
-  }
-  if (
-    event.type === "agent_status" &&
-    isJsonObject(payload) &&
-    typeof payload.status === "string"
-  ) {
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.agent_status",
-      params: {
-        ...base,
-        agentId: base.agentId ?? sessionId,
-        status:
-          payload.status === "error" || payload.status === "stopped"
-            ? payload.status
-            : "idle",
-        ...(agentRunStatusFromPayload(payload.runStatus) !== undefined
-          ? { runStatus: agentRunStatusFromPayload(payload.runStatus) }
-          : {}),
-        ...(typeof payload.turnId === "string"
-          ? { turnId: payload.turnId }
-          : {}),
-        ...(typeof payload.message === "string"
-          ? { message: payload.message }
-          : {}),
-        ...(isJsonObject(payload.budgetHalt)
-          ? { budgetHalt: payload.budgetHalt }
-          : {}),
-        ...(isJsonObject(payload.budgetUsage)
-          ? { budgetUsage: payload.budgetUsage }
-          : {}),
-      },
-    };
-  }
-  if (
-    event.type === "run_terminal" &&
-    isJsonObject(payload) &&
-    typeof payload.status === "string"
-  ) {
-    const terminalStatus = daemonStatusFromRunTerminal(payload.status);
-    const runStatus = agentRunStatusFromRunTerminal(payload.status);
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.agent_status",
-      params: {
-        ...base,
-        agentId: base.agentId ?? sessionId,
-        status: terminalStatus,
-        runStatus,
-        ...(typeof payload.finalMessage === "string"
-          ? { message: payload.finalMessage }
-          : typeof payload.stopReason === "string"
-            ? { message: payload.stopReason }
-            : {}),
-      },
-    };
-  }
-  if (
-    event.type === EVENT_GAP_EVENT &&
-    isJsonObject(payload) &&
-    payload.source === BACKGROUND_RUNNER_GAP_SOURCE &&
-    typeof payload.runId === "string" &&
-    payload.runId.length > 0 &&
-    positiveInteger(payload.retiredCount) > 0
-  ) {
-    const afterSequence = nonNegativeSequence(payload.afterSequence);
-    const firstAvailableSequence = positiveSequence(
-      payload.firstAvailableSequence,
-    );
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.event_gap",
-      params: {
-        ...base,
-        type: EVENT_GAP_EVENT,
-        kind: EVENT_GAP_EVENT,
-        runId: payload.runId,
-        reason: "retention",
-        source: BACKGROUND_RUNNER_GAP_SOURCE,
-        retiredCount: positiveInteger(payload.retiredCount),
-        ...(typeof payload.coordinatesAvailable === "boolean"
-          ? { coordinatesAvailable: payload.coordinatesAvailable }
-          : {}),
-        ...(afterSequence !== undefined ? { afterSequence } : {}),
-        ...(firstAvailableSequence !== undefined
-          ? { firstAvailableSequence }
-          : {}),
-      },
-    };
-  }
-  if (
-    (event.type === "turn_started" ||
-      event.type === "turn_complete" ||
-      event.type === "turn_aborted" ||
-      event.type === "error") &&
-    event.statusProjection !== "session_only" &&
-    isJsonObject(payload)
-  ) {
-    return {
-      jsonrpc: JSON_RPC_VERSION,
-      method: "event.agent_status",
-      params: {
-        ...base,
-        agentId: base.agentId ?? sessionId,
-        status: agentStatusFromEventType(event.type),
-        runStatus: agentRunStatusFromEventType(event.type),
-        ...(typeof payload.turnId === "string"
-          ? { turnId: payload.turnId }
-          : {}),
-        ...(typeof payload.message === "string"
-          ? { message: payload.message }
-          : typeof payload.reason === "string"
-            ? { message: payload.reason }
-            : typeof payload.lastAgentMessage === "string"
-              ? { message: payload.lastAgentMessage }
-              : {}),
-        ...(isJsonObject(payload.budgetHalt)
-          ? { budgetHalt: payload.budgetHalt }
-          : {}),
-        ...(isJsonObject(payload.budgetUsage)
-          ? { budgetUsage: payload.budgetUsage }
-          : {}),
-      },
-    };
-  }
-  return {
-    jsonrpc: JSON_RPC_VERSION,
-    method: "event.session_event",
-    params: {
-      ...base,
-      event: {
-        id: event.id,
-        type: event.type,
-        ...(event.messageId !== undefined
-          ? { messageId: event.messageId }
-          : {}),
-        ...(event.streamId !== undefined ? { streamId: event.streamId } : {}),
-        ...(event.acceptedAt !== undefined
-          ? { acceptedAt: event.acceptedAt }
-          : {}),
-        ...(payload !== undefined ? { payload } : {}),
-      },
-    },
-  };
-}
-
-function eventBaseParams(
-  sessionId: string,
-  agentId: string,
-  event: BackgroundAgentDaemonEvent,
-): {
-  readonly sessionId: string;
-  readonly eventId: string;
-  readonly agentId: string;
-  readonly runId?: string;
-  readonly historyEpoch?: string;
-  readonly sequence?: number;
-  readonly acceptedAt?: string;
-  readonly turnId?: string;
-  readonly clientMessageId?: string;
-  readonly messageId?: string;
-} {
-  return {
-    sessionId,
-    eventId: event.eventId ?? event.id,
-    agentId,
-    ...(event.runId !== undefined ? { runId: event.runId } : {}),
-    ...(event.historyEpoch !== undefined
-      ? { historyEpoch: event.historyEpoch }
-      : {}),
-    ...(event.sequence !== undefined ? { sequence: event.sequence } : {}),
-    ...(event.acceptedAt !== undefined ? { acceptedAt: event.acceptedAt } : {}),
-    ...(event.turnId !== undefined ? { turnId: event.turnId } : {}),
-    ...(event.clientMessageId !== undefined
-      ? { clientMessageId: event.clientMessageId }
-      : {}),
-    ...(event.messageId !== undefined ? { messageId: event.messageId } : {}),
-  };
-}
-
-function agentStatusFromEventType(type: string): DaemonAgentStatus {
-  switch (type) {
-    case "turn_started":
-      return "running";
-    case "error":
-      return "error";
-    case "turn_aborted":
-      return "idle";
-    case "turn_complete":
-    default:
-      return "idle";
-  }
-}
-
-function daemonStatusFromRunTerminal(status: unknown): DaemonAgentStatus {
-  switch (status) {
-    case "completed":
-      return "idle";
-    case "failed":
-    case "unknown_outcome":
-      return "error";
-    case "cancelled":
-    default:
-      return "stopped";
-  }
-}
-
-function agentRunStatusFromRunTerminal(status: unknown): AgentRunStatus {
-  switch (status) {
-    case "completed":
-      return "completed";
-    case "failed":
-    case "unknown_outcome":
-      return "errored";
-    case "cancelled":
-    default:
-      return "stopped";
-  }
-}
-
-function agentRunStatusFromEventType(type: string): AgentRunStatus {
-  switch (type) {
-    case "turn_started":
-      return "running";
-    case "error":
-      return "errored";
-    case "turn_aborted":
-      return "completed";
-    case "turn_complete":
-    default:
-      return "completed";
-  }
-}
-
-async function runRestoredAgentToCompletion(
-  runAgentFn: AgenCRunAgentFunction,
-  opts: {
-    readonly thread: AgentThread;
-    readonly parent: LocalRuntimeBootstrap["session"];
-    readonly registry: ToolRegistry;
-    readonly taskPrompt: string;
-    readonly initialMessages: ReadonlyArray<LLMMessage>;
-    readonly replayToolCalls: readonly AgenCBackgroundAgentReplayToolCall[];
-    readonly currentSessionId?: string;
-    readonly onReplayToolResult?: (
-      result: AgenCBackgroundAgentReplayToolResult,
-    ) => void | Promise<void>;
-    readonly model?: string;
-    readonly onProgress?: (
-      event: RunAgentProgressEvent,
-      thread: AgentThread,
-    ) => void | Promise<void>;
-  },
-): Promise<RunAgentResult> {
-  const replayedMessages = await replayRecoveredToolCalls(opts);
-  const initialMessages =
-    replayedMessages.length === 0
-      ? opts.initialMessages
-      : [...opts.initialMessages, ...replayedMessages];
-  const iter = runAgentFn({
-    live: opts.thread.live,
-    parent: opts.parent,
-    initialMessages,
-    taskPrompt: opts.taskPrompt,
-    ...(opts.model !== undefined ? { model: opts.model } : {}),
-  });
-  // eslint-disable-next-line no-constant-condition
-  while (true) {
-    const step = await iter.next();
-    if (step.done) {
-      return step.value;
-    }
-    await opts.onProgress?.(step.value, opts.thread);
-  }
-}
-
-async function replayRecoveredToolCalls<
-  TThread extends AgentThread | ManagedThread,
->(opts: {
-  readonly thread: TThread;
-  readonly parent: LocalRuntimeBootstrap["session"];
-  readonly registry: ToolRegistry;
-  readonly initialMessages: ReadonlyArray<LLMMessage>;
-  readonly replayToolCalls: readonly AgenCBackgroundAgentReplayToolCall[];
-  readonly currentSessionId?: string;
-  readonly onReplayToolResult?: (
-    result: AgenCBackgroundAgentReplayToolResult,
-  ) => void | Promise<void>;
-  readonly onProgress?: (
-    event: RunAgentProgressEvent,
-    thread: TThread,
-  ) => void | Promise<void>;
-}): Promise<LLMMessage[]> {
-  const messages: LLMMessage[] = [];
-  for (const replay of opts.replayToolCalls) {
-    const args = stringifyReplayToolArguments(replay.args);
-    const registeredTool = opts.registry.tools.find(
-      (tool) => tool.name === replay.toolName,
-    );
-    if (registeredTool?.recoveryCategory !== "idempotent") {
-      if (opts.currentSessionId !== undefined) {
-        await opts.onReplayToolResult?.({
-          sessionId: opts.currentSessionId,
-          callId: replay.callId,
-          toolName: replay.toolName,
-          result: `Recovered tool call ${replay.callId} was not replayed because the current tool registration is missing or not idempotent.`,
-          isError: true,
-          terminalStatus: "poisoned",
-          ...(registeredTool?.recoveryCategory !== undefined
-            ? { recoveryCategory: registeredTool.recoveryCategory }
-            : {}),
-        });
-      }
-      continue;
-    }
-    await opts.onProgress?.(
-      {
-        kind: "tool_call",
-        callId: replay.callId,
-        toolName: replay.toolName,
-        arguments: args,
-        recoveryCategory: "idempotent",
-      },
-      opts.thread,
-    );
-    const result = await dispatchReplayToolCall({
-      registry: opts.registry,
-      session: opts.parent,
-      toolCall: {
-        id: replay.callId,
-        name: replay.toolName,
-        arguments: args,
-      },
-    });
-    if (opts.currentSessionId !== undefined) {
-      await opts.onReplayToolResult?.({
-        sessionId: opts.currentSessionId,
-        callId: replay.callId,
-        toolName: replay.toolName,
-        result: result.content,
-        isError: result.isError === true,
-        terminalStatus: result.isError === true ? "failed" : "completed",
-        recoveryCategory: "idempotent",
-      });
-    }
-    await opts.onProgress?.(
-      {
-        kind: "tool_result",
-        callId: replay.callId,
-        toolName: replay.toolName,
-        result: result.content,
-        isError: result.isError === true,
-      },
-      opts.thread,
-    );
-    if (
-      !hasAssistantToolCall(
-        [...opts.initialMessages, ...messages],
-        replay.callId,
-      )
-    ) {
-      messages.push({
-        role: "assistant",
-        content: "",
-        toolCalls: [
-          {
-            id: replay.callId,
-            name: replay.toolName,
-            arguments: args,
-          },
-        ],
-      });
-    }
-    messages.push({
-      role: "tool",
-      content: frameUntrustedToolResultContent(
-        replay.toolName,
-        result.content,
-        classifyUntrustedToolResult(replay.toolName, registeredTool),
-      ),
-      toolCallId: replay.callId,
-      toolName: replay.toolName,
-    });
-  }
-  return messages;
-}
-
-async function hydrateRecoveredSessionHistory(
-  session: LocalRuntimeBootstrap["session"],
-  params: {
-    readonly initialMessages: ReadonlyArray<LLMMessage>;
-    readonly replayedMessages: ReadonlyArray<LLMMessage>;
-  },
-): Promise<void> {
-  if (
-    params.initialMessages.length === 0 &&
-    params.replayedMessages.length === 0
-  ) {
-    return;
-  }
-  const stateLock = (
-    session as {
-      readonly state?: {
-        with?: (
-          fn: (state: { history?: unknown }) => void | Promise<void>,
-        ) => Promise<void> | void;
-      };
-    }
-  ).state;
-  if (typeof stateLock?.with !== "function") return;
-  await stateLock.with((state) => {
-    const current = Array.isArray(state.history) ? state.history : [];
-    const next =
-      current.length === 0
-        ? [...params.initialMessages, ...params.replayedMessages]
-        : [
-            ...current,
-            ...params.replayedMessages.filter(
-              (message) => !historyContainsRecoveredMessage(current, message),
-            ),
-          ];
-    state.history = next.map(cloneRecoveredLlmMessage);
-  });
-}
-
-function historyContainsRecoveredMessage(
-  history: ReadonlyArray<unknown>,
-  message: LLMMessage,
-): boolean {
-  if (message.role === "assistant" && message.toolCalls !== undefined) {
-    const ids = new Set(message.toolCalls.map((toolCall) => toolCall.id));
-    return history.some((entry) => {
-      if (entry === null || typeof entry !== "object") return false;
-      const toolCalls = (entry as { readonly toolCalls?: unknown }).toolCalls;
-      return (
-        Array.isArray(toolCalls) &&
-        toolCalls.some(
-          (toolCall) =>
-            toolCall !== null &&
-            typeof toolCall === "object" &&
-            ids.has(String((toolCall as { readonly id?: unknown }).id)),
-        )
-      );
-    });
-  }
-  if (message.role === "tool" && typeof message.toolCallId === "string") {
-    return history.some(
-      (entry) =>
-        entry !== null &&
-        typeof entry === "object" &&
-        (entry as { readonly role?: unknown }).role === "tool" &&
-        (entry as { readonly toolCallId?: unknown }).toolCallId ===
-          message.toolCallId,
-    );
-  }
-  return false;
-}
-
-function cloneRecoveredLlmMessage(message: LLMMessage): LLMMessage {
-  return {
-    ...message,
-    content: Array.isArray(message.content)
-      ? message.content.map((part) => ({ ...part }))
-      : message.content,
-    ...(message.toolCalls !== undefined
-      ? {
-          toolCalls: message.toolCalls.map((toolCall) => ({ ...toolCall })),
-        }
-      : {}),
-  };
-}
-
-function hasAssistantToolCall(
-  messages: readonly LLMMessage[],
-  toolCallId: string,
-): boolean {
-  return messages.some(
-    (message) =>
-      message.role === "assistant" &&
-      message.toolCalls?.some((toolCall) => toolCall.id === toolCallId) ===
-        true,
-  );
-}
-
-function stringifyReplayToolArguments(value: JsonValue): string {
-  return JSON.stringify(value);
-}
-
-async function dispatchReplayToolCall(opts: {
-  readonly registry: ToolRegistry;
-  readonly session: LocalRuntimeBootstrap["session"];
-  readonly toolCall: {
-    readonly id: string;
-    readonly name: string;
-    readonly arguments: string;
-  };
-}): Promise<{ readonly content: string; readonly isError?: boolean }> {
-  try {
-    const tool = opts.registry.tools.find(
-      (candidate) => candidate.name === opts.toolCall.name,
-    );
-    if (tool === undefined || typeof tool.execute !== "function") {
-      return {
-        content:
-          "Recovered tool call could not be replayed because the current tool registration is not executable.",
-        isError: true,
-      };
-    }
-    const router = routerFromRegistry(opts.registry);
-    const permissionModeRegistry = opts.session.permissionModeRegistry;
-    const permissionContext = permissionModeRegistry
-      ? buildReplayPermissionContext(opts.session, permissionModeRegistry)
-      : null;
-    const modeChangeRegistry =
-      typeof permissionModeRegistry?.subscribeToModeChange === "function"
-        ? permissionModeRegistry
-        : undefined;
-    return await router.dispatchModelToolCall(opts.toolCall, {
-      session: opts.session as Session,
-      turn: buildReplayTurnContext(opts.session, opts.toolCall.id),
-      tracker: replayNoopTracker,
-      approvalPolicy: "never",
-      sandboxMode: "workspace_write",
-      ...(permissionContext !== null
-        ? {
-            canUseTool: hasPermissionsToUseTool,
-            permissionContext,
-            ...(modeChangeRegistry !== undefined ? { modeChangeRegistry } : {}),
-          }
-        : {}),
-    });
-  } catch (error) {
-    return {
-      content: error instanceof Error ? error.message : String(error),
-      isError: true,
-    };
-  }
-}
-
-const replayNoopTracker = {
-  appendFileDiff: () => {},
-  snapshot: () => [],
-  clear: () => {},
-};
-
-function buildReplayPermissionContext(
-  session: LocalRuntimeBootstrap["session"],
-  permissionModeRegistry: PermissionModeRegistry,
-): ToolEvaluatorContext {
-  const denialTracking =
-    (
-      session as {
-        readonly denialTracking?: ReturnType<typeof freshDenialTracking>;
-      }
-    ).denialTracking ?? freshDenialTracking();
-  return attachContextDefaults({
-    session: session as Session,
-    denialTracking,
-    executionSurface: "headless",
-    getAppState: (): AppStateSnapshot => {
-      const current = permissionModeRegistry.current();
-      return {
-        toolPermissionContext: current,
-        denialTracking,
-        autoModeActive: current.autoModeActive === true,
-      };
-    },
-  });
-}
-
-function buildReplayTurnContext(
-  session: LocalRuntimeBootstrap["session"],
-  subId: string,
-): TurnContext {
-  const sessionRecord = session as {
-    readonly config?: unknown;
-    readonly modelInfo?: unknown;
-    readonly provider?: unknown;
-    readonly cwd?: unknown;
-  };
-  const config = (sessionRecord.config ?? {}) as TurnContext["config"];
-  return {
-    subId,
-    config,
-    configSnapshot: config,
-    modelInfo: (sessionRecord.modelInfo ?? {
-      slug: "background-replay",
-      effectiveContextWindowPercent: 100,
-      contextWindow: 8192,
-      supportedReasoningLevels: [],
-      defaultReasoningSummary: "auto",
-      truncationPolicy: "off",
-      usedFallbackModelMetadata: false,
-    }) as TurnContext["modelInfo"],
-    provider: (sessionRecord.provider ?? {}) as TurnContext["provider"],
-    cwd: typeof sessionRecord.cwd === "string" ? sessionRecord.cwd : "/tmp",
-    realtimeActive: false,
-    modelProviderId: "background-replay",
-    reasoningSummary: "auto",
-    sessionSource: "sdk",
-    dynamicTools: [],
-    depth: 0,
-    toolCallGate: {
-      isReady: () => true,
-      signal: () => {},
-      wait: async () => {},
-    },
-  } as unknown as TurnContext;
-}
-
-function restoredAgentMetadata(
-  params: AgenCBackgroundAgentRestoreParams,
-): AgentMetadata {
-  const metadata = params.metadata;
-  const agentPath =
-    metadataStringField(metadata, "agentPath") ??
-    metadataStringField(metadata, "agent_path") ??
-    joinAgentPath(ROOT_AGENT_PATH, normalizeAgentNameForPath(params.agentId));
-  return normalizeAgentMetadata({
-    agentId: params.agentId,
-    agentPath,
-    ...(metadata?.agentNickname !== undefined
-      ? { agentNickname: metadata.agentNickname }
-      : {}),
-    ...(metadata?.agentRole !== undefined
-      ? { agentRole: metadata.agentRole }
-      : {}),
-    ...(metadata?.agentRoleWorkspaceId !== undefined
-      ? { agentRoleWorkspaceId: metadata.agentRoleWorkspaceId }
-      : {}),
-    ...(metadata?.agentRoleFingerprint !== undefined
-      ? { agentRoleFingerprint: metadata.agentRoleFingerprint }
-      : {}),
-    depth: metadata?.depth ?? 1,
-  });
-}
-
-function metadataStringField(
-  value: JsonObject | undefined,
-  key: string,
-): string | undefined {
-  const field = value?.[key];
-  return typeof field === "string" && field.trim().length > 0
-    ? field.trim()
-    : undefined;
-}
-
-function messageContentToLlmParts(
-  content: MessageContent | undefined,
-): readonly LLMContentPart[] | undefined {
-  if (content === undefined) return undefined;
-  if (typeof content === "string") return [{ type: "text", text: content }];
-  return content.map((part) => {
-    if (part.type === "text") return { type: "text", text: part.text };
-    return {
-      type: "image_url",
-      image_url: { url: part.image_url.url },
-    };
-  });
-}
-
-function metadataStringList(
-  value: JsonObject | undefined,
-  key: string,
-): readonly string[] | undefined {
-  if (
-    value === undefined ||
-    !Object.prototype.hasOwnProperty.call(value, key)
-  ) {
-    return undefined;
-  }
-  const field = value[key];
-  if (!Array.isArray(field)) return undefined;
-  return field.filter(
-    (entry): entry is string =>
-      typeof entry === "string" && entry.trim().length > 0,
-  );
-}
-
-function agentRunStatusFromPayload(value: unknown): AgentRunStatus | undefined {
-  switch (value) {
-    case "pending":
-    case "running":
-    case "working":
-    case "paused":
-    case "blocked":
-    case "suspended":
-    case "completed":
-    case "errored":
-    case "stopped":
-      return value;
-    default:
-      return undefined;
-  }
-}
-
-function toolRequestInputFromPayload(
-  payload: JsonObject,
-): JsonValue | undefined {
-  if (payload.input !== undefined && isJsonValue(payload.input)) {
-    return payload.input;
-  }
-  if (typeof payload.args !== "string") return undefined;
-  try {
-    const parsed: unknown = JSON.parse(payload.args);
-    return isJsonValue(parsed) ? parsed : payload.args;
-  } catch {
-    return payload.args;
-  }
-}
-
-/**
- * Translate session-level events that the PhaseEvent → RunAgentProgressEvent
- * pipeline does not cover into BackgroundAgentDaemonEvents the daemon's
- * notification fan-out can deliver. Currently:
- *
- *   - elicitation/user-input requests (`request_user_input`,
- *     `mcp_elicitation_request`, `mcp_elicitation_complete`)
- *   - durable user transcript messages emitted by runtime turns
- *   - collab-agent lifecycle events emitted by `spawn_agent`,
- *     `wait_agent`, `send_message`, and `close_agent`
- *   - runtime warnings emitted before or during a turn
- *   - streaming tool progress chunks (`tool_progress`)
- *   - extended-thinking + reasoning-summary streaming events
- *     (`assistant_thinking_block_start`/`delta`/`block_stop`,
- *     `agent_thinking`)
- *
- * The bridge subscribes to `session.eventLog` (live; writes to the rollout
- * are a separate downstream consumer) and forwards a translated event into
- * the same `#emitOrBufferEvent` pipeline that PhaseEvents use.
- */
-const COLLAB_AGENT_SESSION_EVENT_TYPES: ReadonlySet<string> = new Set([
-  "collab_agent_spawn_begin",
-  "collab_agent_spawn_end",
-  "collab_agent_status",
-  "collab_agent_interaction_begin",
-  "collab_agent_interaction_end",
-  "collab_waiting_begin",
-  "collab_waiting_end",
-  "collab_close_begin",
-  "collab_close_end",
-]);
-
-/**
- * Core run events whose live representation must be the exact canonical
- * Session.EventLog record. Returning null for an unsequenced record is
- * intentional: synthesizing an identity here would make live delivery
- * impossible to reconcile with run.replay after reconnect.
- */
-const CANONICAL_CORE_SESSION_EVENT_TYPES: ReadonlySet<string> = new Set([
-  "history_cleared",
-  "transcript_epoch",
-  "agent_message_delta",
-  "tool_call_started",
-  "tool_call_completed",
-  "turn_started",
-  "turn_complete",
-  "turn_aborted",
-  "warning",
-  "error",
-  "stream_error",
-  "effect_intent",
-  "effect_result",
-  "effect_unknown_outcome",
-  "effect_review_resolved",
-  "request_permissions",
-  "permission_decision",
-  "execution_admission",
-  "artifact_intent",
-  "artifact_committed",
-  "recovery_decision",
-  "run_terminal",
-  "run_reopened",
-  "run_suspended",
-  "run_resumed",
-  "run_runtime_settings_changed",
-  "run_cancel_requested",
-]);
-
-export function daemonEventFromUnboundSessionEvent(event: {
-  readonly eventId?: unknown;
-  readonly id?: unknown;
-  readonly seq?: unknown;
-  readonly msg?: {
-    readonly type?: unknown;
-    readonly payload?: unknown;
-  };
-}): BackgroundAgentDaemonEvent | null {
-  const type = event.msg?.type;
-  const payload = event.msg?.payload;
-  const id =
-    typeof event.id === "string" && event.id.length > 0
-      ? event.id
-      : typeof type === "string"
-        ? type
-        : "elicitation";
-  const sequence =
-    typeof event.seq === "number" &&
-    Number.isSafeInteger(event.seq) &&
-    event.seq > 0
-      ? event.seq
-      : undefined;
-  const eventId =
-    typeof event.eventId === "string" && event.eventId.length > 0
-      ? event.eventId
-      : sequence !== undefined
-        ? `legacy-event:${sequence}:${id}`
-        : id;
-  if (
-    typeof type === "string" &&
-    CANONICAL_CORE_SESSION_EVENT_TYPES.has(type)
-  ) {
-    if (sequence === undefined || !isJsonObject(payload)) return null;
-    return { id, eventId, sequence, type, payload };
-  }
-  if (
-    type === "request_user_input" &&
-    isJsonObject(payload) &&
-    typeof payload.callId === "string" &&
-    typeof payload.turnId === "string" &&
-    Array.isArray(payload.questions)
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload: {
-        callId: payload.callId,
-        requestId:
-          typeof payload.requestId === "string"
-            ? payload.requestId
-            : payload.callId,
-        turnId: payload.turnId,
-        questions: jsonObjectArray(payload.questions),
-        ...(isJsonObject(payload.clientAction)
-          ? { clientAction: payload.clientAction }
-          : {}),
-      },
-    };
-  }
-  if (
-    type === "mcp_elicitation_request" &&
-    isJsonObject(payload) &&
-    typeof payload.serverName === "string" &&
-    (typeof payload.requestId === "string" ||
-      typeof payload.requestId === "number") &&
-    typeof payload.turnId === "string" &&
-    isJsonObject(payload.request)
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload: {
-        serverName: payload.serverName,
-        requestId: payload.requestId,
-        turnId: payload.turnId,
-        request: payload.request,
-      },
-    };
-  }
-  if (
-    type === "mcp_elicitation_complete" &&
-    isJsonObject(payload) &&
-    typeof payload.serverName === "string" &&
-    typeof payload.elicitationId === "string"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload: {
-        serverName: payload.serverName,
-        elicitationId: payload.elicitationId,
-      },
-    };
-  }
-  if (
-    type === "user_message" &&
-    isJsonObject(payload) &&
-    payload.message !== undefined
-  ) {
-    const messageId =
-      typeof payload.messageId === "string" ? payload.messageId : undefined;
-    const streamId =
-      typeof payload.streamId === "string" ? payload.streamId : undefined;
-    const acceptedAt =
-      typeof payload.acceptedAt === "string" ? payload.acceptedAt : undefined;
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      ...(messageId !== undefined ? { messageId } : {}),
-      ...(streamId !== undefined ? { streamId } : {}),
-      ...(acceptedAt !== undefined ? { acceptedAt } : {}),
-      payload: {
-        message: payload.message,
-        ...(messageId !== undefined ? { messageId } : {}),
-        ...(streamId !== undefined ? { streamId } : {}),
-        ...(acceptedAt !== undefined ? { acceptedAt } : {}),
-        ...(typeof payload.displayText === "string"
-          ? { displayText: payload.displayText }
-          : {}),
-        ...(Array.isArray(payload.images)
-          ? { images: stringArray(payload.images) }
-          : {}),
-        ...(typeof payload.queuedCommandUuid === "string"
-          ? { queuedCommandUuid: payload.queuedCommandUuid }
-          : {}),
-      },
-    };
-  }
-  if (
-    typeof type === "string" &&
-    COLLAB_AGENT_SESSION_EVENT_TYPES.has(type) &&
-    isJsonObject(payload) &&
-    typeof payload.callId === "string"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload,
-    };
-  }
-  // Provider token accounting for the latest model call. Telemetry like
-  // collab_agent_status: clients drive live context gauges off it, and
-  // without forwarding it a UI can only estimate context from transcript
-  // characters — blind to tool output, which is where context really goes.
-  if (
-    type === "token_count" &&
-    isJsonObject(payload) &&
-    typeof payload.promptTokens === "number" &&
-    typeof payload.totalTokens === "number"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload,
-    };
-  }
-  if (
-    type === "tool_progress" &&
-    isJsonObject(payload) &&
-    typeof payload.callId === "string" &&
-    typeof payload.toolName === "string"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload,
-    };
-  }
-  // Assistant message-complete boundaries. `agent_message` is emitted per
-  // completed assistant message segment (run-turn) and persisted to the
-  // rollout, but the live wire only carried `agent_message_delta` — so a
-  // daemon-attached TUI accumulated the deltas of CONSECUTIVE messages
-  // into one streaming buffer with no separator ("…subagents.No M1-named
-  // files…"). The reducer's agent_message case is what closes a segment
-  // (pushes the completed row, clears the buffer); forward it live.
-  if (
-    type === "agent_message" &&
-    isJsonObject(payload) &&
-    typeof payload.message === "string"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload,
-    };
-  }
-  // Per-stream provider usage. Emitted via `session.emit` from
-  // `phases/stream-model.ts` (T6 #119) and persisted to the rollout, but
-  // never bridged live — so a daemon-attached TUI had no usage source at
-  // all: its synthesized assistant messages carry zero usage and the
-  // workbench ctx% read 0 for the whole session. Forward the payload
-  // verbatim; the TUI reducer derives `latestUsage` from it.
-  if (type === "token_count" && isJsonObject(payload)) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload,
-    };
-  }
-  // Streaming tool-argument events (input_json_delta family). Same live
-  // bridge gap as the thinking events below: providers that stream tool
-  // arguments (grok function_call_arguments deltas, Messages-API
-  // input_json_delta) emit these through stream-model, and the TUI needs
-  // them for in-flight tool rendering, the spinner's cumulative token
-  // estimate, and stream-liveness accounting.
-  if (
-    type === "tool_input_block_start" &&
-    isJsonObject(payload) &&
-    typeof payload.callId === "string" &&
-    typeof payload.index === "number"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload,
-    };
-  }
-  if (
-    type === "tool_input_delta" &&
-    isJsonObject(payload) &&
-    typeof payload.callId === "string" &&
-    typeof payload.index === "number" &&
-    typeof payload.partialJson === "string"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload,
-    };
-  }
-  // Extended-thinking + reasoning-summary events. These are emitted via
-  // `session.emit` from `phases/stream-model.ts` and persisted to the
-  // rollout, but the live notification path needs an explicit bridge:
-  // the PhaseEvent → RunAgentProgressEvent → BackgroundAgentDaemonEvent
-  // pipeline at `phaseEventToProgressEvent` does not carry them, so the
-  // TUI's `event.session_event` catch-all never sees them without this.
-  if (
-    type === "assistant_thinking_block_start" &&
-    isJsonObject(payload) &&
-    typeof payload.index === "number"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload: {
-        index: payload.index,
-        redacted: payload.redacted === true,
-        ...(typeof payload.kind === "string" ? { kind: payload.kind } : {}),
-      },
-    };
-  }
-  if (
-    type === "assistant_thinking_delta" &&
-    isJsonObject(payload) &&
-    typeof payload.delta === "string" &&
-    typeof payload.index === "number"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload: {
-        delta: payload.delta,
-        index: payload.index,
-        ...(typeof payload.kind === "string" ? { kind: payload.kind } : {}),
-      },
-    };
-  }
-  if (
-    type === "assistant_thinking_block_stop" &&
-    isJsonObject(payload) &&
-    typeof payload.index === "number"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload: {
-        index: payload.index,
-        ...(typeof payload.kind === "string" ? { kind: payload.kind } : {}),
-      },
-    };
-  }
-  if (
-    type === "agent_thinking" &&
-    isJsonObject(payload) &&
-    typeof payload.text === "string"
-  ) {
-    return {
-      id,
-      eventId,
-      ...(sequence !== undefined ? { sequence } : {}),
-      type,
-      payload: {
-        text: payload.text,
-        ...(payload.redacted === true ? { redacted: true } : {}),
-        ...(typeof payload.kind === "string" ? { kind: payload.kind } : {}),
-      },
-    };
-  }
-  return null;
-}
-
-function jsonObjectArray(value: readonly unknown[]): JsonObject[] {
-  return value.filter(isJsonObject);
-}
-
-function isJsonObject(value: unknown): value is JsonObject {
-  return isRecord(value);
-}
-
-function isJsonValue(value: unknown): value is JsonValue {
-  if (value === null) return true;
-  switch (typeof value) {
-    case "string":
-    case "number":
-    case "boolean":
-      return true;
-    case "object":
-      if (Array.isArray(value)) return value.every(isJsonValue);
-      return Object.values(value).every(
-        (item) => item === undefined || isJsonValue(item),
-      );
-    default:
-      return false;
-  }
-}
-
-function stringArray(value: unknown): readonly string[] {
-  return Array.isArray(value)
-    ? value.filter((item): item is string => typeof item === "string")
-    : [];
-}
-
-function isToolRecoveryCategory(
-  value: unknown,
-): value is "idempotent" | "side-effecting" | "interactive" {
-  return (
-    value === "idempotent" ||
-    value === "side-effecting" ||
-    value === "interactive"
-  );
-}
-
-interface ManagedTokenUsageShape {
-  readonly inputTokens: number;
-  readonly outputTokens: number;
-  readonly totalTokens: number;
-}
-
-export function managedTokenUsage(
-  thread: Pick<ManagedThread, "totalTokenUsage">,
-): ManagedTokenUsageShape {
-  const usage = thread.totalTokenUsage?.();
-  if (typeof usage !== "object" || usage === null) {
-    return { inputTokens: 0, outputTokens: 0, totalTokens: 0 };
-  }
-  const u = usage as Record<string, unknown>;
-  // Two shapes reach this seam: run-agent's live counter uses
-  // inputTokens/outputTokens, while a daemon session's cross-turn
-  // accumulator (stream-model.ts, the TokenUsageInfo port) uses
-  // promptTokens/completionTokens. Reading only the former zeroed
-  // input/output in every session.snapshot (totalTokens matched both
-  // shapes, which is why the bug shipped as {0, 0, N}).
-  const field = (...names: readonly string[]): number => {
-    for (const name of names) {
-      const value = u[name];
-      if (typeof value === "number" && Number.isFinite(value)) return value;
-    }
-    return 0;
-  };
-  const inputTokens = field("inputTokens", "promptTokens");
-  const outputTokens = field("outputTokens", "completionTokens");
-  const totalTokens = field("totalTokens");
-  return {
-    inputTokens,
-    outputTokens,
-    totalTokens: totalTokens > 0 ? totalTokens : inputTokens + outputTokens,
-  };
-}
-
-// Translate Session PhaseEvents only for runner-local status/tool bookkeeping.
-// Live delivery is owned by the canonical Session.EventLog bridge above; using
-// this phase shape for delivery would invent competing IDs without sequences.
-// Exported as a test seam: the stop-reason mapping decides whether a turn
-// outcome ends the turn or the whole run.
-export function phaseEventToProgressEvent(
-  event: import("../phases/events.js").PhaseEvent,
-): RunAgentProgressEvent | null {
-  switch (event.type) {
-    case "turn_start":
-      return null;
-    case "history_cleared":
-      return null;
-    case "queued_command":
-      return null;
-    case "assistant_text":
-      return {
-        kind: "message",
-        message: { role: "assistant", content: event.content },
-      };
-    case "tool_call":
-      return {
-        kind: "tool_call",
-        callId: event.toolCall.id,
-        toolName: event.toolCall.name,
-        arguments: event.toolCall.arguments,
-      };
-    case "tool_result":
-      return {
-        kind: "tool_result",
-        callId: event.toolCall.id,
-        toolName: event.toolCall.name,
-        result: event.result.content,
-        isError: event.result.isError === true,
-      };
-    case "turn_complete": {
-      const turnId = `turn-${event.stopReason}-${event.content.length}-${
-        event.usage?.totalTokens ?? 0
-      }`;
-      if (event.stopReason === "cancelled") {
-        return {
-          kind: "turn_interrupted",
-          reason: "cancelled",
-          turnId,
-        };
-      }
-      if (event.stopReason === "error") {
-        // A turn that errored, most often a provider call the network dropped
-        // after dispatch, is a per-turn outcome too. Mapping it to run_error
-        // flipped the agent to status=error and every later prompt was
-        // refused with "no longer running": one connection error ended a
-        // session whose earlier turns had all settled. The turn ends with the
-        // failure spelled out; the agent stays idle and takes the next prompt.
-        const reason = event.error?.message?.trim() || "turn errored";
-        const sentence = /[.!?]$/u.test(reason) ? reason : `${reason}.`;
-        return {
-          kind: "turn_complete",
-          turnId,
-          toolCallCount: 0,
-          finalMessage: `Turn failed: ${sentence} Send a new prompt to retry.`,
-        };
-      }
-      // Bounded stops — the backstop, a turn cap, the cost cap, a
-      // compact skip/throw, and a request-scoped Editor failure — are
-      // per-TURN outcomes, not run deaths. Mapping them to run_error
-      // bricked the whole session: the user saw "no longer running
-      // (status: error)" and could never prompt again after one bad
-      // turn. The turn ends honestly with its message; the session
-      // stays available for the next prompt, exactly like "completed".
-      const boundedStopFallback: Partial<Record<string, string>> = {
-        max_turns: "Turn capped: iteration limit hit; send a new prompt to continue.",
-        max_budget_usd: "Turn capped: cost ceiling hit; send a new prompt to continue.",
-        no_progress: "Turn halted by the progress backstop; send a new prompt to continue.",
-        compact_failed:
-          "Turn stopped: compaction could not shrink the context; send a new prompt to continue.",
-        editor_request_failed:
-          "Editor request stopped safely; send a new prompt to continue.",
-      };
-      const boundedFallback = boundedStopFallback[event.stopReason];
-      if (boundedFallback !== undefined) {
-        const preferredMessage =
-          (event.stopReason === "compact_failed" ||
-            event.stopReason === "editor_request_failed") &&
-          event.error instanceof Error &&
-          event.error.message.length > 0
-            ? event.error.message
-            : undefined;
-        return {
-          kind: "turn_complete",
-          turnId,
-          toolCallCount: 0,
-          finalMessage:
-            preferredMessage ??
-            (event.content.length > 0 ? event.content : boundedFallback),
-        };
-      }
-      // "completed" | "empty_response" — a per-turn completion. Emit
-      // turn_complete (NOT run_complete — the session continues across
-      // turns; run_complete would trigger cleanup).
-      return {
-        kind: "turn_complete",
-        turnId,
-        toolCallCount: 0,
-        ...(event.content.length > 0 ? { finalMessage: event.content } : {}),
-      };
-    }
-  }
-}
-
-function canonicalSessionEventFromRecoveredProgress(
-  progress: RunAgentProgressEvent,
-): Event | null {
-  if (progress.kind === "tool_call") {
-    return {
-      id: `recovery-tool-start:${progress.callId}`,
-      msg: {
-        type: "tool_call_started",
-        payload: {
-          callId: progress.callId,
-          toolName: progress.toolName,
-          args: progress.arguments ?? "{}",
-        },
-      },
-    };
-  }
-  if (progress.kind === "tool_result") {
-    return {
-      id: `recovery-tool-result:${progress.callId}`,
-      msg: {
-        type: "tool_call_completed",
-        payload: {
-          callId: progress.callId,
-          result: progress.result,
-          isError: progress.isError === true,
-          metadata: {
-            toolName: progress.toolName,
-            recovered: true,
-          },
-        },
-      },
-    };
-  }
-  return null;
-}
-
-type TerminalThreadStatus = Extract<
-  ThreadAgentStatus,
-  { readonly status: "completed" | "errored" | "shutdown" | "not_found" }
->;
-
-interface CapturedRuntimeSettingsOptions {
-  readonly profile?: string;
-  readonly permissionContext?: ToolPermissionContext;
-}
-
-function runtimeWorkspaceRoot(bootstrap: LocalRuntimeBootstrap): string {
-  const broker = bootstrap.session.services.sandboxExecutionBroker;
-  if (!(broker instanceof SandboxExecutionBroker)) {
-    throw new Error(
-      "canonical runtime settings require the live sandbox execution broker cwd",
-    );
-  }
-  const cwd = broker.cwd;
-  if (typeof cwd !== "string" || cwd.trim().length === 0) {
-    throw new Error("live sandbox execution broker cwd is unavailable");
-  }
-  return canonicalizeBypassPermissionsCwd(cwd);
-}
-
-function supportsCanonicalRuntimeSettings(
-  active: ActiveBackgroundAgent,
-): boolean {
-  return (
-    typeof active.bootstrap.session.permissionModeRegistry
-      .installBeforeUpdateHook === "function" &&
-    typeof active.bootstrap.session.permissionModeRegistry
-      .installPublicationCoordinator === "function" &&
-    active.bootstrap.session.services.sandboxExecutionBroker instanceof
-      SandboxExecutionBroker &&
-    active.bootstrap.configuredExecutionAuthority !== undefined &&
-    typeof active.bootstrap.prepareConfiguredExecutionAuthority ===
-      "function" &&
-    typeof active.bootstrap.rolloutStore.recordRunRuntimeSettingsEvent ===
-      "function"
-  );
-}
-
-function requireCanonicalRuntimeSettingsSupport(
-  active: ActiveBackgroundAgent,
-  runId: string,
-): void {
-  if (!supportsCanonicalRuntimeSettings(active)) {
-    throw new Error(
-      `run ${runId} requires a canonical permission registry and durable runtime-settings journal`,
-    );
-  }
-}
-
-function failClosedDaemonRuntimeAuthority(
-  active: ActiveBackgroundAgent,
-  error: unknown,
-  options: {
-    readonly brokerReason: string;
-    readonly abortReason: Parameters<Session["abortTerminal"]>[0];
-    readonly abortFailureMessage: string;
-  },
-): never {
-  const session = active.bootstrap.session;
-  const broker = session.services.sandboxExecutionBroker;
-  if (broker instanceof SandboxExecutionBroker) {
-    broker.closeAfterLifecycleAuthorityFailure(options.brokerReason);
-  }
-  active.ingressClosed = true;
-  try {
-    session.abortTerminal(options.abortReason);
-  } catch (abortError) {
-    throw new AggregateError([error, abortError], options.abortFailureMessage, {
-      cause: error,
-    });
-  }
-  throw error;
-}
-
-function installDaemonPermissionAuthorityCoordinator(
-  bootstrap: LocalRuntimeBootstrap,
-  owner: () => ActiveBackgroundAgent | undefined,
-): () => void {
-  const session = bootstrap.session;
-  const registry = session.permissionModeRegistry;
-  const broker = session.services.sandboxExecutionBroker;
-  if (!(broker instanceof SandboxExecutionBroker)) {
-    throw new Error(
-      "daemon session requires the canonical sandbox execution broker",
-    );
-  }
-  if (
-    bootstrap.configuredExecutionAuthority === undefined ||
-    typeof bootstrap.prepareConfiguredExecutionAuthority !== "function"
-  ) {
-    throw new Error(
-      "daemon session requires a configured execution-authority snapshot",
-    );
-  }
-
-  return registry.installPublicationCoordinator(
-    async (
-      next,
-      _current,
-      metadata,
-      publication: PermissionContextPublication,
-    ) => {
-      const stagedConfiguredAuthority =
-        configuredExecutionAuthorityFromPublicationMetadata(metadata);
-      const preparedConfigReload =
-        preparedConfigStoreReloadFromPublicationMetadata(metadata);
-      const preparedMcpAuthorityRefresh =
-        preparedMcpAuthorityRefreshFromPublicationMetadata(metadata);
-      const authority = executionAuthorityForPermissionContext(
-        stagedConfiguredAuthority?.authority ??
-          bootstrap.configuredExecutionAuthority,
-        next,
-        session.services.runtimeOptions
-          ?.dangerouslyBypassApprovalsAndSandbox === true,
-      );
-      let previousConfiguration: Session["sessionConfiguration"] | undefined;
-      let configurationWriteStarted = false;
-      let stagedConfiguredAuthorityCommitted = false;
-      let preparedConfigReloadCommitted = false;
-      let preparedMcpAuthorityRefreshStarted = false;
-      let authorityTransitionCompleted = false;
-      try {
-        await transitionSandboxExecutionBrokerAuthority(
-          broker,
-          sandboxExecutionBrokerAuthorityFromSessionAuthority(
-            authority,
-            broker.cwd,
-          ),
-          {
-            commit: async () => {
-              configurationWriteStarted = true;
-              await session.state.with((state) => {
-                previousConfiguration = state.sessionConfiguration;
-                state.sessionConfiguration = applySessionExecutionAuthority(
-                  state.sessionConfiguration,
-                  authority,
-                );
-              });
-              preparedConfigReload?.commit();
-              preparedConfigReloadCommitted =
-                preparedConfigReload !== undefined;
-              stagedConfiguredAuthority?.commit();
-              stagedConfiguredAuthorityCommitted =
-                stagedConfiguredAuthority !== undefined;
-              await publication.commit();
-              preparedConfigReload?.publish(
-                COORDINATED_CONFIG_STORE_PUBLICATION,
-              );
-              preparedMcpAuthorityRefreshStarted =
-                preparedMcpAuthorityRefresh !== undefined;
-              preparedMcpAuthorityRefresh?.start();
-              await preparedMcpAuthorityRefresh?.waitUntilDeferred();
-            },
-            rollback: async () => {
-              const rollbackErrors: unknown[] = [];
-              try {
-                await publication.rollback();
-              } catch (error) {
-                rollbackErrors.push(error);
-              }
-              if (stagedConfiguredAuthorityCommitted) {
-                try {
-                  stagedConfiguredAuthority?.rollback();
-                  stagedConfiguredAuthorityCommitted = false;
-                } catch (error) {
-                  rollbackErrors.push(error);
-                }
-              }
-              if (preparedConfigReloadCommitted) {
-                try {
-                  preparedConfigReload?.rollback();
-                  preparedConfigReloadCommitted = false;
-                } catch (error) {
-                  rollbackErrors.push(error);
-                }
-              }
-              if (
-                configurationWriteStarted &&
-                previousConfiguration !== undefined
-              ) {
-                try {
-                  await session.state.with((state) => {
-                    state.sessionConfiguration = previousConfiguration!;
-                  });
-                } catch (error) {
-                  rollbackErrors.push(error);
-                }
-              }
-              if (rollbackErrors.length > 0) {
-                throw new AggregateError(
-                  rollbackErrors,
-                  "daemon permission authority rollback incomplete",
-                );
-              }
-            },
-          },
-        );
-        authorityTransitionCompleted = true;
-        await preparedMcpAuthorityRefresh?.settle();
-        preparedConfigReload?.settle();
-      } catch (error) {
-        const cleanupErrors: unknown[] = [];
-        if (!authorityTransitionCompleted) {
-          try {
-            await publication.rollback();
-          } catch (rollbackError) {
-            cleanupErrors.push(rollbackError);
-          }
-          if (
-            preparedConfigReload !== undefined &&
-            !preparedConfigReload.settled
-          ) {
-            if (preparedConfigReload.state !== "rolled_back") {
-              try {
-                preparedConfigReload.rollback();
-              } catch (rollbackError) {
-                cleanupErrors.push(rollbackError);
-              }
-            }
-            try {
-              preparedConfigReload.settle();
-            } catch (settleError) {
-              cleanupErrors.push(settleError);
-            }
-          }
-        } else {
-          if (
-            preparedConfigReload !== undefined &&
-            !preparedConfigReload.settled
-          ) {
-            try {
-              preparedConfigReload.settle();
-            } catch (settleError) {
-              cleanupErrors.push(settleError);
-            }
-          }
-          if (!broker.isClosedAfterLifecycleAuthorityFailure()) {
-            broker.closeAfterLifecycleAuthorityFailure(
-              "daemon permission authority failed after canonical publication",
-            );
-          }
-        }
-        const failure =
-          cleanupErrors.length === 0
-            ? error
-            : new AggregateError(
-                [error, ...cleanupErrors],
-                "daemon permission authority cleanup was incomplete",
-                { cause: error },
-              );
-        if (
-          (cleanupErrors.length > 0 || preparedMcpAuthorityRefreshStarted) &&
-          !broker.isClosedAfterLifecycleAuthorityFailure()
-        ) {
-          broker.closeAfterLifecycleAuthorityFailure(
-            preparedMcpAuthorityRefreshStarted
-              ? "daemon permission authority failed after canonical publication"
-              : "daemon permission authority cleanup was incomplete",
-          );
-        }
-        if (!broker.isClosedAfterLifecycleAuthorityFailure()) {
-          if (
-            failure instanceof AggregateError &&
-            failure.errors.length === 1
-          ) {
-            throw failure.errors[0];
-          }
-          throw failure;
-        }
-        const active = owner();
-        if (active !== undefined) active.ingressClosed = true;
-        const terminalFailure =
-          failure instanceof AggregateError && failure.errors.length === 1
-            ? failure.errors[0]
-            : failure;
-        try {
-          session.abortTerminal("permission_authority_failure");
-        } catch (abortError) {
-          throw new AggregateError(
-            [terminalFailure, abortError],
-            "daemon permission authority failed and session abort was incomplete",
-            { cause: terminalFailure },
-          );
-        }
-        throw terminalFailure;
-      }
-    },
-  );
-}
-
-function preparedConfigStoreReloadFromPublicationMetadata(
-  metadata: unknown,
-): PreparedConfigStoreReload | undefined {
-  if (!isRecord(metadata)) return undefined;
-  const prepared = metadata.preparedConfigReload;
-  if (!isRecord(prepared)) return undefined;
-  if (
-    !isRecord(prepared.authority) ||
-    typeof prepared.commit !== "function" ||
-    typeof prepared.publish !== "function" ||
-    typeof prepared.rollback !== "function" ||
-    typeof prepared.settle !== "function"
-  ) {
-    throw new Error("prepared config reload publication metadata is invalid");
-  }
-  return prepared as unknown as PreparedConfigStoreReload;
-}
-
-interface PreparedMcpAuthorityRefresh {
-  readonly result: McpRefreshResult | undefined;
-  start(): void;
-  waitUntilDeferred(): Promise<void>;
-  settle(): Promise<void>;
-}
-
-function prepareMcpAuthorityRefresh(
-  session: Session,
-): PreparedMcpAuthorityRefresh | undefined {
-  const manager = session.services.mcpManager;
-  const refresh = manager?.refreshFromAuthority;
-  if (manager === undefined || refresh === undefined) return undefined;
-  let task: Promise<McpRefreshResult> | undefined;
-  let result: McpRefreshResult | undefined;
-  let deferred = false;
-  let readySettled = false;
-  let resolveReady!: () => void;
-  let rejectReady!: (error: unknown) => void;
-  const ready = new Promise<void>((resolve, reject) => {
-    resolveReady = resolve;
-    rejectReady = reject;
-  });
-  void ready.catch(() => undefined);
-  const markDeferred = (): void => {
-    if (readySettled) return;
-    deferred = true;
-    readySettled = true;
-    resolveReady();
-  };
-  const failReady = (error: unknown): void => {
-    if (readySettled) return;
-    readySettled = true;
-    rejectReady(error);
-  };
-  return Object.freeze({
-    get result() {
-      return result;
-    },
-    start: () => {
-      if (task !== undefined) {
-        throw new Error("MCP authority refresh was started more than once");
-      }
-      try {
-        task = Promise.resolve(
-          refresh.call(manager, {
-            onSandboxRefreshDeferred: markDeferred,
-          }),
-        );
-      } catch (error) {
-        task = Promise.reject(error);
-      }
-      void task.then(() => {
-        if (!deferred) {
-          failReady(
-            new Error(
-              "MCP authority refresh completed before sandbox deferral was proven",
-            ),
-          );
-        }
-      }, failReady);
-      void task.catch(() => undefined);
-    },
-    waitUntilDeferred: () => ready,
-    settle: async () => {
-      if (task === undefined) {
-        throw new Error("MCP authority refresh was not started");
-      }
-      result = await task;
-    },
-  });
-}
-
-function preparedMcpAuthorityRefreshFromPublicationMetadata(
-  metadata: unknown,
-): PreparedMcpAuthorityRefresh | undefined {
-  if (!isRecord(metadata)) return undefined;
-  const prepared = metadata.preparedMcpAuthorityRefresh;
-  if (!isRecord(prepared)) return undefined;
-  if (
-    typeof prepared.start !== "function" ||
-    typeof prepared.waitUntilDeferred !== "function" ||
-    typeof prepared.settle !== "function"
-  ) {
-    throw new Error("prepared MCP refresh publication metadata is invalid");
-  }
-  return prepared as unknown as PreparedMcpAuthorityRefresh;
-}
-
-function configuredExecutionAuthorityFromPublicationMetadata(
-  metadata: unknown,
-): PreparedConfiguredExecutionAuthority | undefined {
-  if (!isRecord(metadata)) return undefined;
-  const prepared = metadata.configuredExecutionAuthority;
-  if (!isRecord(prepared)) return undefined;
-  if (
-    !isRecord(prepared.authority) ||
-    typeof prepared.commit !== "function" ||
-    typeof prepared.rollback !== "function"
-  ) {
-    throw new Error(
-      "configured execution authority publication metadata is invalid",
-    );
-  }
-  return prepared as unknown as PreparedConfiguredExecutionAuthority;
-}
-
-function captureRuntimeSettings(
-  active: ActiveBackgroundAgent,
-  options: CapturedRuntimeSettingsOptions = {},
-): RunRuntimeSettingsSnapshot {
-  const { bootstrap } = active;
-  const session = bootstrap.session;
-  const workspaceRoot = runtimeWorkspaceRoot(bootstrap);
-  const permission =
-    options.permissionContext ?? session.permissionModeRegistry.current();
-  if (permission.mode === "bubble") {
-    throw new Error("root daemon runtime settings cannot persist bubble mode");
-  }
-  const prePlanMode =
-    permission.mode === "plan"
-      ? permission.prePlanMode === undefined ||
-        permission.prePlanMode === "bubble"
-        ? "default"
-        : permission.prePlanMode
-      : null;
-  const bypassTransitionCritical =
-    permission.mode === "bypassPermissions" ||
-    prePlanMode === "bypassPermissions";
-  const hasSessionExactBypassConsent =
-    permission.bypassPermissionsAcceptedIn?.includes(workspaceRoot) === true;
-  if (bypassTransitionCritical && !hasSessionExactBypassConsent) {
-    throw new Error(
-      `cannot persist bypass permission authority without exact workspace consent: ${workspaceRoot}`,
-    );
-  }
-  let hasDurableExactBypassConsent = false;
-  try {
-    hasDurableExactBypassConsent =
-      loadBypassPermissionsConsent(
-        bootstrap.configStore.stateRepository,
-        workspaceRoot,
-        { reload: true },
-      )[0] === workspaceRoot;
-  } catch {
-    // A failed state refresh cannot add authority. Session authority remains
-    // usable for an already-active transition, but is not widened here.
-  }
-  const bypassDisabledByPolicy =
-    permission.bypassPermissionsModeDisabledByPolicy === true;
-  const hasExactBypassConsent =
-    !bypassDisabledByPolicy &&
-    (hasSessionExactBypassConsent || hasDurableExactBypassConsent);
-  const bypassPermissionsModeAvailable =
-    !bypassDisabledByPolicy &&
-    (permission.isBypassPermissionsModeAvailable === true ||
-      hasExactBypassConsent);
-
-  const pending = session.pendingProviderSwitch;
-  const selection = readSessionSelection(session, { includePending: true });
-  const configuration = (
-    session as Session & {
-      readonly sessionConfiguration?: Session["sessionConfiguration"];
-    }
-  ).sessionConfiguration;
-  const reasoningEffort = normalizeRuntimeSetting(
-    configuration?.collaborationMode.reasoningEffort,
-    RUN_RUNTIME_REASONING_EFFORTS,
-    "reasoning effort",
-  );
-  const modelVerbosity = normalizeRuntimeSetting(
-    configuration?.modelVerbosity,
-    RUN_RUNTIME_MODEL_VERBOSITIES,
-    "model verbosity",
-  );
-  const serviceTier = normalizeRuntimeSetting(
-    configuration?.serviceTier,
-    RUN_RUNTIME_SERVICE_TIERS,
-    "service tier",
-  );
-  return cloneFrozenRuntimeSettingsSnapshot({
-    permissionMode: permission.mode,
-    prePlanMode,
-    autoModeActive: permission.autoModeActive === true,
-    autoModeAvailable: permission.isAutoModeAvailable === true,
-    bypassPermissionsModeAvailable,
-    bypassPermissionsWorkspace: bypassTransitionCritical ? workspaceRoot : null,
-    bypassPermissionsConsentWorkspace: hasExactBypassConsent
-      ? workspaceRoot
-      : null,
-    model: selection.model,
-    provider: selection.provider,
-    profile:
-      options.profile ??
-      pending?.profile ??
-      active.runtimeSettings?.profile ??
-      null,
-    reasoningEffort,
-    modelVerbosity,
-    serviceTier,
-    hooksDisabled: session.services?.hooksRuntime?.isDisabled() === true,
-  });
-}
-
-function normalizeRuntimeSetting<const T extends readonly string[]>(
-  value: unknown,
-  accepted: T,
-  label: string,
-): T[number] | null {
-  if (value === undefined || value === null) return null;
-  if (
-    typeof value !== "string" ||
-    !(accepted as readonly string[]).includes(value)
-  ) {
-    throw new Error(`cannot persist unsupported ${label}: ${String(value)}`);
-  }
-  return value as T[number];
-}
-
-function installRuntimeSettingsPreCommit(
-  active: ActiveBackgroundAgent,
-  runId: string,
-): () => void {
-  const registry = active.bootstrap.session.permissionModeRegistry;
-  requireCanonicalRuntimeSettingsSupport(active, runId);
-  return registry.installBeforeUpdateHook(async (next, current, metadata) => {
-    const release = await acquireRuntimeSettingsMutation(active);
-    try {
-      if (active.ingressClosed === true) {
-        throw new Error(`run ${runId} permission ingress is closed`);
-      }
-      if (active.runtimeSettingsEventId === undefined) {
-        const baseline = captureRuntimeSettings(active, {
-          permissionContext: current,
-        });
-        commitDurableRuntimeSettingsChange(active, runId, baseline, "initial");
-      }
-      const previousSettings = active.runtimeSettings!;
-      const nextSettings = captureRuntimeSettings(active, {
-        permissionContext: next,
-      });
-      if (
-        active.runtimeSettings !== undefined &&
-        stableStringify(active.runtimeSettings) ===
-          stableStringify(nextSettings)
-      ) {
-        release();
-        return undefined;
-      }
-      const prepared = prepareDurableRuntimeSettingsChange(
-        active,
-        runId,
-        nextSettings,
-        runtimeSettingsCommitMetadata(metadata)?.reason ??
-          "permission_mode_changed",
-        runtimeSettingsCommitMetadata(metadata)?.rollbackOfSettingsEventId ??
-          null,
-      );
-      return {
-        commit: () => {
-          prepared.finalize();
-        },
-        rollback: () => {
-          if (active.runtimeSettingsEventId === prepared.eventId) {
-            compensateRuntimeSettingsChange(
-              active,
-              runId,
-              previousSettings,
-              prepared.eventId,
-            );
-            return;
-          }
-          compensatePreparedRuntimeSettingsChange(
-            active,
-            runId,
-            previousSettings,
-            prepared,
-          );
-        },
-        settle: release,
-      };
-    } catch (error) {
-      release();
-      throw error;
-    }
-  });
-}
-
-function runtimeSettingsCommitMetadata(metadata: unknown):
-  | {
-      readonly reason: RunRuntimeSettingsChangeReason;
-      readonly rollbackOfSettingsEventId: string | null;
-    }
-  | undefined {
-  if (
-    metadata === null ||
-    typeof metadata !== "object" ||
-    !("runtimeSettings" in metadata)
-  ) {
-    return undefined;
-  }
-  const value = (metadata as { readonly runtimeSettings?: unknown })
-    .runtimeSettings;
-  if (value === null || typeof value !== "object") return undefined;
-  const candidate = value as {
-    readonly reason?: unknown;
-    readonly rollbackOfSettingsEventId?: unknown;
-  };
-  if (
-    typeof candidate.reason !== "string" ||
-    !RUN_RUNTIME_SETTINGS_CHANGE_REASONS.includes(candidate.reason as never) ||
-    (candidate.rollbackOfSettingsEventId !== null &&
-      typeof candidate.rollbackOfSettingsEventId !== "string")
-  ) {
-    throw new Error("invalid permission runtime-settings commit metadata");
-  }
-  return {
-    reason: candidate.reason as RunRuntimeSettingsChangeReason,
-    rollbackOfSettingsEventId: candidate.rollbackOfSettingsEventId,
-  };
-}
-
-async function withRuntimeSettingsMutation<T>(
-  active: ActiveBackgroundAgent,
-  mutate: () => Promise<T>,
-): Promise<T> {
-  const result = active.runtimeSettingsMutationQueue.then(mutate);
-  active.runtimeSettingsMutationQueue = result.then(
-    () => {},
-    () => {},
-  );
-  return result;
-}
-
-async function acquireRuntimeSettingsMutation(
-  active: ActiveBackgroundAgent,
-): Promise<() => void> {
-  const previous = active.runtimeSettingsMutationQueue;
-  let release!: () => void;
-  const held = new Promise<void>((resolve) => {
-    release = resolve;
-  });
-  const acquired = previous.then(
-    () => {},
-    () => {},
-  );
-  active.runtimeSettingsMutationQueue = acquired.then(() => held);
-  await acquired;
-  return release;
-}
-
-function ensureInitialRuntimeSettings(
-  active: ActiveBackgroundAgent,
-  runId: string,
-): RunRuntimeSettingsSnapshot {
-  if (
-    active.runtimeSettings !== undefined &&
-    active.runtimeSettingsEventId !== undefined
-  ) {
-    return active.runtimeSettings;
-  }
-  const baseline = captureRuntimeSettings(active);
-  commitDurableRuntimeSettingsChange(active, runId, baseline, "initial");
-  return active.runtimeSettings!;
-}
-
-function failClosedRuntimeSettingsAuthority(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  error: unknown,
-): never {
-  return failClosedDaemonRuntimeAuthority(active, error, {
-    brokerReason: "daemon runtime-settings authority is ambiguous",
-    abortReason: "permission_authority_failure",
-    abortFailureMessage: `run ${runId} runtime-settings authority failed and session abort was incomplete`,
-  });
-}
-
-function compensateRuntimeSettingsChange(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  previous: RunRuntimeSettingsSnapshot,
-  failedSettingsEventId: string,
-): void {
-  try {
-    if (active.runtimeSettingsEventId !== failedSettingsEventId) {
-      throw new Error(
-        `run ${runId} settings compensation no longer follows ${failedSettingsEventId}`,
-      );
-    }
-    commitDurableRuntimeSettingsChange(
-      active,
-      runId,
-      previous,
-      "compensating_rollback",
-      failedSettingsEventId,
-    );
-  } catch (error) {
-    failClosedRuntimeSettingsAuthority(active, runId, error);
-  }
-}
-
-function compensatePreparedRuntimeSettingsChange(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  previous: RunRuntimeSettingsSnapshot,
-  failed: PreparedRuntimeSettingsChange,
-): void {
-  try {
-    if (active.runtimeSettingsEventId !== failed.previousSettingsEventId) {
-      throw new Error(
-        `run ${runId} settings compensation no longer follows ${failed.eventId}`,
-      );
-    }
-    const compensation = prepareDurableRuntimeSettingsChange(
-      active,
-      runId,
-      previous,
-      "compensating_rollback",
-      failed.eventId,
-      failed.eventId,
-    );
-    projectDurableRuntimeSettingsEvent(active, failed.event);
-    compensation.finalize();
-  } catch (error) {
-    projectDurableRuntimeSettingsEvent(active, failed.event);
-    failClosedRuntimeSettingsAuthority(active, runId, error);
-  }
-}
-
-async function applyRestoredRuntimeSettings(
-  bootstrap: LocalRuntimeBootstrap,
-  settings: RunRuntimeSettingsSnapshot,
-): Promise<RunRuntimeSettingsSnapshot> {
-  const workspaceRoot = canonicalizeBypassPermissionsCwd(
-    runtimeWorkspaceRoot(bootstrap),
-  );
-  assertValidRuntimeSettingsSnapshot(settings, workspaceRoot);
-  const bypassTransitionCritical =
-    settings.permissionMode === "bypassPermissions" ||
-    (settings.permissionMode === "plan" &&
-      settings.prePlanMode === "bypassPermissions");
-  if (
-    (bypassTransitionCritical &&
-      settings.bypassPermissionsWorkspace !== workspaceRoot) ||
-    (!bypassTransitionCritical && settings.bypassPermissionsWorkspace !== null)
-  ) {
-    throw new Error(
-      "canonical bypass permission workspace does not match restored workspace",
-    );
-  }
-  const session = bootstrap.session;
-  const registry = session.permissionModeRegistry;
-  const current = registry.current();
-  const bypassDisabledByPolicy =
-    current.bypassPermissionsModeDisabledByPolicy === true;
-  const [persistedBypassConsent] = loadBypassPermissionsConsent(
-    bootstrap.configStore.stateRepository,
-    workspaceRoot,
-    { reload: true },
-  );
-  const hasCurrentDurableBypassConsent =
-    persistedBypassConsent === workspaceRoot;
-  const autoModeAvailable =
-    settings.autoModeAvailable && current.isAutoModeAvailable === true;
-  const retainedConsent =
-    !bypassDisabledByPolicy &&
-    settings.bypassPermissionsModeAvailable &&
-    settings.bypassPermissionsConsentWorkspace === workspaceRoot &&
-    hasCurrentDurableBypassConsent;
-  const bypassModeAvailable =
-    !bypassDisabledByPolicy &&
-    (current.isBypassPermissionsModeAvailable === true || retainedConsent);
-  let transitionContext: ToolPermissionContext = {
-    ...current,
-    isAutoModeAvailable: autoModeAvailable,
-    isBypassPermissionsModeAvailable: bypassModeAvailable,
-    bypassPermissionsAcceptedIn: retainedConsent ? [workspaceRoot] : [],
-  };
-  if (bypassTransitionCritical) {
-    if (bypassDisabledByPolicy) {
-      throw new Error(
-        "restored bypass permission mode is disabled by managed policy",
-      );
-    }
-    if (!hasCurrentDurableBypassConsent) {
-      throw new Error(
-        "restored bypass permission mode requires persisted exact-cwd consent",
-      );
-    }
-    transitionContext = authorizeBypassPermissionsConsent(
-      {
-        ...transitionContext,
-        isBypassPermissionsModeAvailable: false,
-        bypassPermissionsAcceptedIn: [],
-      },
-      persistedBypassConsent,
-    );
-  }
-  let transitioned = runWithCurrentRuntimeSession(session, () =>
-    settings.permissionMode === "bypassPermissions"
-      ? transitionPermissionMode(
-          transitionContext.mode,
-          settings.permissionMode,
-          transitionContext,
-          { workspacePath: workspaceRoot },
-        )
-      : transitionPermissionMode(
-          transitionContext.mode,
-          settings.permissionMode,
-          transitionContext,
-        ),
-  );
-  if ("error" in transitioned) {
-    throw new Error(
-      "restored bypass permission mode lacks exact canonical workspace consent",
-    );
-  }
-  const bypassAccepted =
-    bypassTransitionCritical || retainedConsent ? [workspaceRoot] : [];
-  transitioned = {
-    ...transitioned,
-    mode: settings.permissionMode,
-    ...(settings.permissionMode === "plan"
-      ? { prePlanMode: settings.prePlanMode ?? "default" }
-      : { prePlanMode: undefined }),
-    autoModeActive: settings.autoModeActive,
-    isAutoModeAvailable: autoModeAvailable,
-    isBypassPermissionsModeAvailable: bypassModeAvailable,
-    bypassPermissionsAcceptedIn: bypassAccepted,
-  };
-  await registry.update(transitioned);
-
-  const liveSelection = readSessionSelection(session);
-  if (
-    liveSelection.provider !== settings.provider ||
-    liveSelection.model !== settings.model ||
-    settings.profile !== null
-  ) {
-    session.setPendingProviderSwitch({
-      provider: settings.provider,
-      model: settings.model,
-      ...(settings.profile !== null ? { profile: settings.profile } : {}),
-    });
-  }
-  await session.state.with((state) => {
-    const configuration = state.sessionConfiguration;
-    state.sessionConfiguration = {
-      ...configuration,
-      collaborationMode: {
-        ...configuration.collaborationMode,
-        ...(settings.reasoningEffort !== null
-          ? { reasoningEffort: settings.reasoningEffort }
-          : { reasoningEffort: undefined }),
-      } as typeof configuration.collaborationMode,
-      ...(settings.modelVerbosity !== null
-        ? { modelVerbosity: settings.modelVerbosity }
-        : { modelVerbosity: undefined }),
-      ...(settings.serviceTier !== null
-        ? { serviceTier: settings.serviceTier }
-        : { serviceTier: undefined }),
-    };
-  });
-  session.services?.hooksRuntime?.setDisabled(settings.hooksDisabled);
-  return cloneFrozenRuntimeSettingsSnapshot({
-    ...settings,
-    autoModeAvailable,
-    bypassPermissionsModeAvailable: bypassModeAvailable,
-    bypassPermissionsConsentWorkspace: retainedConsent ? workspaceRoot : null,
-  });
-}
-
-function currentCanonicalRuntimeStateFromRollout(
-  bootstrap: LocalRuntimeBootstrap,
-  runId: string,
-): {
-  readonly pendingStartupActivationResumeEventId?: string;
-  readonly runtimeSettings?: RunRuntimeSettingsSnapshot;
-  readonly runtimeSettingsEventId?: string;
-} {
-  const epoch = currentRunEpochFromRollout(bootstrap, runId);
-  let pendingStartupActivationResumeEventId: string | undefined;
-  let runtimeSettings: RunRuntimeSettingsSnapshot | undefined;
-  let runtimeSettingsEventId: string | undefined;
-  for (const item of bootstrap.rolloutStore.readAll()) {
-    if (item.type !== "event_msg") continue;
-    const event = item.payload;
-    const payload = event.msg.payload as { runId?: unknown; epoch?: unknown };
-    if (payload.runId !== runId) continue;
-    if (
-      event.msg.type === "run_runtime_settings_changed" &&
-      typeof payload.epoch === "number" &&
-      payload.epoch <= epoch
-    ) {
-      runtimeSettings = runtimeSettingsSnapshotFromCanonicalEvent(event);
-      runtimeSettingsEventId = canonicalEventId(event);
-    }
-    if (payload.epoch !== epoch) continue;
-    if (event.msg.type === "run_resumed") {
-      pendingStartupActivationResumeEventId = canonicalEventId(event);
-    } else if (event.msg.type === "run_startup_activated") {
-      if (
-        event.msg.payload.resumeEventId ===
-        pendingStartupActivationResumeEventId
-      ) {
-        pendingStartupActivationResumeEventId = undefined;
-      }
-    } else if (
-      event.msg.type === "run_suspended" ||
-      event.msg.type === "run_terminal"
-    ) {
-      pendingStartupActivationResumeEventId = undefined;
-    }
-  }
-  return {
-    ...(pendingStartupActivationResumeEventId !== undefined
-      ? { pendingStartupActivationResumeEventId }
-      : {}),
-    ...(runtimeSettings !== undefined ? { runtimeSettings } : {}),
-    ...(runtimeSettingsEventId !== undefined ? { runtimeSettingsEventId } : {}),
-  };
-}
-
-function runtimeSettingsSnapshotFromCanonicalEvent(
-  event: Event,
-): RunRuntimeSettingsSnapshot {
-  if (event.msg.type !== "run_runtime_settings_changed") {
-    throw new Error("expected canonical runtime settings event");
-  }
-  const payload = event.msg.payload;
-  return cloneFrozenRuntimeSettingsSnapshot({
-    permissionMode: payload.permissionMode,
-    prePlanMode: payload.prePlanMode,
-    autoModeActive: payload.autoModeActive,
-    autoModeAvailable: payload.autoModeAvailable,
-    bypassPermissionsModeAvailable: payload.bypassPermissionsModeAvailable,
-    bypassPermissionsWorkspace: payload.bypassPermissionsWorkspace,
-    bypassPermissionsConsentWorkspace:
-      payload.bypassPermissionsConsentWorkspace,
-    model: payload.model,
-    provider: payload.provider,
-    profile: payload.profile,
-    reasoningEffort: payload.reasoningEffort,
-    modelVerbosity: payload.modelVerbosity,
-    serviceTier: payload.serviceTier,
-    hooksDisabled: payload.hooksDisabled,
-  });
-}
-
-function commitDurableRuntimeSettingsChange(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  settings: RunRuntimeSettingsSnapshot,
-  reason: RunRuntimeSettingsChangeReason,
-  rollbackOfSettingsEventId: string | null = null,
-): void {
-  prepareDurableRuntimeSettingsChange(
-    active,
-    runId,
-    settings,
-    reason,
-    rollbackOfSettingsEventId,
-  ).finalize();
-}
-
-interface PreparedRuntimeSettingsChange {
-  readonly event: Event;
-  readonly eventId: string;
-  readonly previousSettingsEventId: string | null;
-  finalize(): void;
-}
-
-function prepareDurableRuntimeSettingsChange(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  settings: RunRuntimeSettingsSnapshot,
-  reason: RunRuntimeSettingsChangeReason,
-  rollbackOfSettingsEventId: string | null = null,
-  preparedPredecessorEventId?: string,
-): PreparedRuntimeSettingsChange {
-  if (!supportsCanonicalRuntimeSettings(active)) {
-    throw new Error(
-      `run ${runId} cannot change runtime settings without canonical journal support`,
-    );
-  }
-  const canonicalSettings = cloneFrozenRuntimeSettingsSnapshot(settings);
-  assertValidRuntimeSettingsSnapshot(
-    canonicalSettings,
-    runtimeWorkspaceRoot(active.bootstrap),
-  );
-  const epoch = active.runEpoch;
-  const previousSettingsEventId =
-    preparedPredecessorEventId ?? active.runtimeSettingsEventId ?? null;
-  if (
-    preparedPredecessorEventId !== undefined &&
-    (reason !== "compensating_rollback" ||
-      rollbackOfSettingsEventId !== preparedPredecessorEventId)
-  ) {
-    throw new Error(
-      `run ${runId} prepared predecessor is only valid for its compensation`,
-    );
-  }
-  if (previousSettingsEventId === null && reason !== "initial") {
-    throw new Error(
-      `run ${runId} must establish initial runtime settings before ${reason}`,
-    );
-  }
-  if (previousSettingsEventId !== null && reason === "initial") {
-    throw new Error(`run ${runId} already has initial runtime settings`);
-  }
-  const eventId = `run-runtime-settings:${runId}:${epoch}:${randomUUID()}`;
-  const changedAt = new Date().toISOString();
-  const acceptCommitted = (proveDurable: boolean): Event | undefined => {
-    const matches = active.bootstrap.rolloutStore
-      .readAll()
-      .flatMap((item) =>
-        item.type === "event_msg" &&
-        (item.payload.eventId === eventId || item.payload.id === eventId)
-          ? [item.payload]
-          : [],
-      );
-    if (matches.length === 0) return undefined;
-    if (matches.length !== 1) {
-      throw new Error(`runtime settings ${eventId} has duplicate evidence`);
-    }
-    const event = matches[0]!;
-    if (
-      event.id !== eventId ||
-      event.eventId !== eventId ||
-      positiveSequence(event.seq) === undefined ||
-      event.msg.type !== "run_runtime_settings_changed" ||
-      event.msg.payload.runId !== runId ||
-      event.msg.payload.epoch !== epoch ||
-      event.msg.payload.previousSettingsEventId !== previousSettingsEventId ||
-      event.msg.payload.rollbackOfSettingsEventId !==
-        rollbackOfSettingsEventId ||
-      event.msg.payload.reason !== reason ||
-      event.msg.payload.changedAt !== changedAt ||
-      stableStringify(runtimeSettingsSnapshotFromCanonicalEvent(event)) !==
-        stableStringify(canonicalSettings)
-    ) {
-      throw new Error(`runtime settings ${eventId} has conflicting evidence`);
-    }
-    if (proveDurable) {
-      active.bootstrap.rolloutStore.syncCanonicalTail();
-      return acceptCommitted(false);
-    }
-    return event;
-  };
-  let event: Event;
-  let publish: () => Event;
-  try {
-    const candidate = {
-      eventId,
-      id: eventId,
-      msg: {
-        type: "run_runtime_settings_changed",
-        payload: {
-          runId,
-          epoch,
-          previousSettingsEventId,
-          rollbackOfSettingsEventId,
-          reason,
-          changedAt,
-          ...canonicalSettings,
-        },
-      },
-    } satisfies Event;
-    const prepared = active.bootstrap.session.prepareEmit(candidate);
-    event = prepared.event;
-    publish = prepared.publish;
-  } catch (error) {
-    let recovered: Event | undefined;
-    try {
-      recovered = acceptCommitted(true);
-    } catch (evidenceError) {
-      failClosedRuntimeSettingsAuthority(
-        active,
-        runId,
-        new AggregateError(
-          [error, evidenceError],
-          `runtime settings ${eventId} preparation failed after an ambiguous canonical append`,
-          { cause: error },
-        ),
-      );
-    }
-    if (recovered === undefined) throw error;
-    event = recovered;
-    publish = () => active.bootstrap.session.publishPreparedEvent(recovered);
-  }
-  if (event.eventId !== eventId || positiveSequence(event.seq) === undefined) {
-    throw new Error(`runtime settings ${eventId} lacks canonical coordinates`);
-  }
-  let finalized = false;
-  return {
-    event,
-    eventId,
-    previousSettingsEventId,
-    finalize: () => {
-      if (finalized) return;
-      finalized = true;
-      active.runtimeSettings = canonicalSettings;
-      active.runtimeSettingsEventId = eventId;
-      try {
-        publish();
-      } catch (publishError) {
-        let failure: unknown = publishError;
-        try {
-          const committed = acceptCommitted(true);
-          if (committed === undefined) {
-            throw new Error(
-              `runtime settings ${eventId} publication failed without canonical evidence`,
-            );
-          }
-          projectDurableRuntimeSettingsEvent(active, committed);
-        } catch (evidenceError) {
-          failure = new AggregateError(
-            [publishError, evidenceError],
-            `runtime settings ${eventId} publication failed and canonical evidence could not be proved`,
-            { cause: publishError },
-          );
-        }
-        failClosedRuntimeSettingsAuthority(active, runId, failure);
-      }
-      projectDurableRuntimeSettingsEvent(active, event);
-    },
-  };
-}
-
-function projectDurableRuntimeSettingsEvent(
-  active: ActiveBackgroundAgent,
-  event: Event,
-): void {
-  try {
-    active.bootstrap.rolloutStore.recordRunRuntimeSettingsEvent(event);
-  } catch {
-    // Canonical fsync evidence is authoritative; SQLite is rebuildable.
-  }
-}
-
-function assertValidRuntimeSettingsSnapshot(
-  settings: RunRuntimeSettingsSnapshot,
-  workspaceRoot: string,
-): void {
-  const bounded = (value: string, maxBytes: number): boolean =>
-    value.trim().length > 0 && Buffer.byteLength(value, "utf8") <= maxBytes;
-  const nullableBounded = (value: string | null, maxBytes: number): boolean =>
-    value === null || bounded(value, maxBytes);
-  const bypassTransitionCritical =
-    settings.permissionMode === "bypassPermissions" ||
-    settings.prePlanMode === "bypassPermissions";
-  const hasExactBypassConsent =
-    settings.bypassPermissionsConsentWorkspace === workspaceRoot;
-  let providerModelIsCanonical = false;
-  try {
-    const selection = mergeProviderModelLayer(
-      {},
-      {
-        model_provider: settings.provider,
-        model: settings.model,
-      },
-    );
-    providerModelIsCanonical =
-      selection.model_provider === settings.provider &&
-      selection.model === settings.model;
-  } catch {
-    providerModelIsCanonical = false;
-  }
-  if (
-    !RUN_RUNTIME_PERMISSION_MODES.includes(settings.permissionMode) ||
-    (settings.prePlanMode !== null &&
-      !RUN_RUNTIME_PERMISSION_MODES.includes(settings.prePlanMode)) ||
-    (settings.permissionMode === "plan"
-      ? settings.prePlanMode === null || settings.prePlanMode === "plan"
-      : settings.prePlanMode !== null) ||
-    (settings.permissionMode === "auto"
-      ? settings.autoModeActive !== true
-      : settings.permissionMode !== "plan" &&
-        settings.autoModeActive !== false) ||
-    typeof settings.autoModeAvailable !== "boolean" ||
-    (settings.autoModeActive && !settings.autoModeAvailable) ||
-    typeof settings.bypassPermissionsModeAvailable !== "boolean" ||
-    (settings.bypassPermissionsConsentWorkspace !== null &&
-      !hasExactBypassConsent) ||
-    (hasExactBypassConsent && !settings.bypassPermissionsModeAvailable) ||
-    (bypassTransitionCritical
-      ? settings.bypassPermissionsWorkspace !== workspaceRoot ||
-        !settings.bypassPermissionsModeAvailable ||
-        !hasExactBypassConsent
-      : settings.bypassPermissionsWorkspace !== null) ||
-    !bounded(settings.model, 1_024) ||
-    !bounded(settings.provider, 256) ||
-    !providerModelIsCanonical ||
-    !nullableBounded(settings.profile, 256) ||
-    (settings.reasoningEffort !== null &&
-      !RUN_RUNTIME_REASONING_EFFORTS.includes(settings.reasoningEffort)) ||
-    (settings.modelVerbosity !== null &&
-      !RUN_RUNTIME_MODEL_VERBOSITIES.includes(settings.modelVerbosity)) ||
-    (settings.serviceTier !== null &&
-      !RUN_RUNTIME_SERVICE_TIERS.includes(settings.serviceTier)) ||
-    typeof settings.hooksDisabled !== "boolean"
-  ) {
-    throw new Error("runtime settings snapshot is not canonically valid");
-  }
-}
-
-function commitDurableRunStartupActivation(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  activatedAt: string,
-): void {
-  const resumeEventId = active.pendingStartupActivationResumeEventId;
-  if (resumeEventId === undefined) return;
-  const exactActivatedAt =
-    active.pendingStartupActivationActivatedAt ?? activatedAt;
-  active.pendingStartupActivationActivatedAt = exactActivatedAt;
-  const epoch = active.runEpoch;
-  const resumeHash = createHash("sha256")
-    .update(resumeEventId, "utf8")
-    .digest("hex")
-    .slice(0, 32);
-  const eventId = `run-startup-activated:${runId}:${epoch}:${resumeHash}`;
-  const acceptCommitted = (proveDurable: boolean): Event | undefined => {
-    const matches = active.bootstrap.rolloutStore
-      .readAll()
-      .flatMap((item) =>
-        item.type === "event_msg" &&
-        (item.payload.eventId === eventId || item.payload.id === eventId)
-          ? [item.payload]
-          : [],
-      );
-    if (matches.length === 0) return undefined;
-    if (matches.length !== 1) {
-      throw new Error(`startup activation ${eventId} has duplicate evidence`);
-    }
-    const event = matches[0]!;
-    if (
-      event.id !== eventId ||
-      event.eventId !== eventId ||
-      positiveSequence(event.seq) === undefined ||
-      event.msg.type !== "run_startup_activated" ||
-      event.msg.payload.runId !== runId ||
-      event.msg.payload.epoch !== epoch ||
-      event.msg.payload.resumeEventId !== resumeEventId ||
-      event.msg.payload.activatedAt !== exactActivatedAt
-    ) {
-      throw new Error(`startup activation ${eventId} has conflicting evidence`);
-    }
-    if (proveDurable) {
-      active.bootstrap.rolloutStore.syncCanonicalTail();
-      return acceptCommitted(false);
-    }
-    return event;
-  };
-  let event: Event;
-  try {
-    event = active.bootstrap.session.emit({
-      eventId,
-      id: eventId,
-      msg: {
-        type: "run_startup_activated",
-        payload: {
-          runId,
-          epoch,
-          resumeEventId,
-          activatedAt: exactActivatedAt,
-        },
-      },
-    });
-  } catch (error) {
-    const recovered = acceptCommitted(true);
-    if (recovered === undefined) throw error;
-    event = recovered;
-  }
-  if (event.eventId !== eventId || positiveSequence(event.seq) === undefined) {
-    throw new Error(
-      `startup activation ${eventId} lacks canonical coordinates`,
-    );
-  }
-  try {
-    active.bootstrap.rolloutStore.recordRunStartupActivationEvent(event);
-  } catch {
-    // Canonical fsync evidence is authoritative; SQLite is rebuildable.
-  }
-  active.pendingStartupActivationResumeEventId = undefined;
-  active.pendingStartupActivationActivatedAt = undefined;
-}
-
-function awaitTerminalStatus(
-  thread: ManagedThread,
-): Promise<TerminalThreadStatus> {
-  return new Promise((resolve) => {
-    let settledSynchronously = false;
-    let unsubscribe = (): void => {};
-    const listener = (status: ThreadAgentStatus): void => {
-      if (
-        status.status === "completed" ||
-        status.status === "errored" ||
-        status.status === "shutdown" ||
-        status.status === "not_found"
-      ) {
-        settledSynchronously = true;
-        unsubscribe();
-        resolve(status);
-      }
-    };
-    unsubscribe = thread.subscribeStatus(listener);
-    // ManagedThread subscriptions publish their current value immediately.
-    // If it was already terminal, the callback ran before the real
-    // unsubscribe function was assigned.
-    if (settledSynchronously) unsubscribe();
-  });
-}
-
-function commitDurableRunCancellationRequest(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  reason: string,
-): void {
-  const requestedAt = active.pendingTerminal?.finishedAt ?? active.lastActiveAt;
-  const existing = active.cancellationRequest;
-  if (existing !== undefined) {
-    if (existing.reason !== reason || existing.requestedAt !== requestedAt) {
-      throw new Error(`run ${runId} has conflicting cancellation intent`);
-    }
-    return;
-  }
-  const eventId = `run-cancel-request:${runId}:${active.runEpoch}`;
-  const acceptCommitted = (proveDurable = false): boolean => {
-    const matches = active.bootstrap.rolloutStore.readAll().flatMap((item) => {
-      if (item.type !== "event_msg") return [];
-      const event = item.payload;
-      if (
-        event.eventId !== eventId &&
-        event.id !== eventId &&
-        !(
-          event.msg.type === "run_cancel_requested" &&
-          event.msg.payload.runId === runId &&
-          event.msg.payload.epoch === active.runEpoch
-        )
-      ) {
-        return [];
-      }
-      return [event];
-    });
-    if (matches.length === 0) return false;
-    if (matches.length !== 1) {
-      throw new Error(
-        `run cancellation request ${eventId} has duplicate canonical evidence`,
-      );
-    }
-    const event = matches[0]!;
-    const sequence = positiveSequence(event.seq);
-    if (
-      event.id !== eventId ||
-      event.eventId !== eventId ||
-      sequence === undefined ||
-      event.msg.type !== "run_cancel_requested" ||
-      event.msg.payload.runId !== runId ||
-      event.msg.payload.epoch !== active.runEpoch ||
-      event.msg.payload.reason !== reason ||
-      event.msg.payload.requestedAt !== requestedAt
-    ) {
-      throw new Error(
-        `run cancellation request ${eventId} has conflicting canonical evidence`,
-      );
-    }
-    if (proveDurable) {
-      active.bootstrap.rolloutStore.syncCanonicalTail();
-      return acceptCommitted(false);
-    }
-    active.cancellationRequest = {
-      eventId,
-      sequence,
-      reason,
-      requestedAt,
-    };
-    return true;
-  };
-  if (acceptCommitted(true)) return;
-  try {
-    const event = active.bootstrap.session.emit({
-      eventId,
-      id: eventId,
-      msg: {
-        type: "run_cancel_requested",
-        payload: {
-          runId,
-          epoch: active.runEpoch,
-          reason,
-          requestedAt,
-        },
-      },
-    });
-    const sequence = positiveSequence(event.seq);
-    if (
-      event.id !== eventId ||
-      event.eventId !== eventId ||
-      sequence === undefined
-    ) {
-      throw new Error(
-        `run cancellation request ${eventId} has no canonical coordinates`,
-      );
-    }
-    active.cancellationRequest = {
-      eventId,
-      sequence,
-      reason,
-      requestedAt,
-    };
-  } catch (error) {
-    // Session.emit may fail after append+fsync at the publish failpoint. The
-    // deterministic identity makes retry safe only when the bytes on disk are
-    // exactly the requested cancellation evidence.
-    if (!acceptCommitted(true)) throw error;
-  }
-}
-
-function commitDurableRunTerminal(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  result: RunTerminalResult,
-): AgenCBackgroundAgentTerminalSnapshot {
-  if (active.terminal !== undefined) return active.terminal;
-  const epoch = active.runEpoch;
-  const session = active.bootstrap.session;
-  const lastSequenceBeforeTerminal =
-    positiveSequence(session.eventLog.lastSeq) ?? null;
-  const eventId = `run-terminal:${runId}:${epoch}`;
-  const event = session.emit({
-    eventId,
-    id: eventId,
-    msg: {
-      type: "run_terminal",
-      payload: {
-        runId,
-        epoch,
-        status: result.status,
-        exitCode: result.exitCode,
-        stopReason: result.stopReason,
-        finalMessage: result.finalMessage,
-        usage: result.usage,
-        lastSequenceBeforeTerminal,
-        finishedAt: result.finishedAt,
-      },
-    },
-  });
-  const sequence = positiveSequence(event.seq);
-  if (
-    event.id !== eventId ||
-    event.eventId !== eventId ||
-    sequence === undefined
-  ) {
-    throw new Error(
-      `run_terminal ${eventId} was not assigned its canonical id and positive sequence`,
-    );
-  }
-  const terminal: AgenCBackgroundAgentTerminalSnapshot = {
-    openedAt: active.startedAt,
-    epoch,
-    eventId: event.eventId,
-    rolloutPath: active.bootstrap.rolloutStore.rolloutPath,
-    result: {
-      ...result,
-      lastSequence: sequence,
-    },
-  };
-  active.terminal = terminal;
-  return terminal;
-}
-
-function commitDurableRunSuspension(
-  active: ActiveBackgroundAgent,
-  runId: string,
-): AgenCBackgroundAgentSuspensionSnapshot {
-  if (active.suspension !== undefined) return active.suspension;
-  const pending = active.pendingSuspension;
-  if (pending === undefined) {
-    throw new Error(`run ${runId} has no daemon suspension pending`);
-  }
-  const epoch = active.runEpoch;
-  const eventId = pending.eventId;
-
-  const acceptCommitted = (
-    proveDurable = false,
-  ): AgenCBackgroundAgentSuspensionSnapshot | undefined => {
-    const matches = active.bootstrap.rolloutStore.readAll().flatMap((item) => {
-      if (item.type !== "event_msg") return [];
-      const event = item.payload;
-      if (event.eventId !== eventId && event.id !== eventId) {
-        return [];
-      }
-      if (
-        event.msg.type !== "run_suspended" ||
-        event.msg.payload.runId !== runId ||
-        event.msg.payload.epoch !== epoch
-      ) {
-        return [];
-      }
-      return [event];
-    });
-    if (matches.length === 0) return undefined;
-    if (matches.length !== 1) {
-      throw new Error(
-        `run suspension ${eventId} has duplicate canonical evidence`,
-      );
-    }
-    const event = matches[0]!;
-    const sequence = positiveSequence(event.seq);
-    if (
-      event.id !== eventId ||
-      event.eventId !== eventId ||
-      sequence === undefined ||
-      event.msg.type !== "run_suspended" ||
-      event.msg.payload.reason !== pending.reason ||
-      event.msg.payload.suspendedAt !== pending.suspendedAt
-    ) {
-      throw new Error(
-        `run suspension ${eventId} has conflicting canonical evidence`,
-      );
-    }
-    if (proveDurable) {
-      active.bootstrap.rolloutStore.syncCanonicalTail();
-      return acceptCommitted(false);
-    }
-    const suspension: AgenCBackgroundAgentSuspensionSnapshot = {
-      openedAt: active.startedAt,
-      epoch,
-      eventId,
-      sequence,
-      rolloutPath: active.bootstrap.rolloutStore.rolloutPath,
-      reason: pending.reason,
-      suspendedAt: pending.suspendedAt,
-    };
-    active.suspension = suspension;
-    try {
-      active.bootstrap.rolloutStore.recordRunSuspensionEvent(event);
-    } catch {
-      // SQLite is a rebuildable projection. Canonical fsync evidence remains
-      // authoritative and startup recovery will replay this boundary.
-    }
-    return suspension;
-  };
-
-  const committed = acceptCommitted(true);
-  if (committed !== undefined) return committed;
-  try {
-    const event = active.bootstrap.session.emit({
-      eventId,
-      id: eventId,
-      msg: {
-        type: "run_suspended",
-        payload: {
-          runId,
-          epoch,
-          reason: pending.reason,
-          suspendedAt: pending.suspendedAt,
-        },
-      },
-    });
-    const sequence = positiveSequence(event.seq);
-    if (
-      event.id !== eventId ||
-      event.eventId !== eventId ||
-      sequence === undefined
-    ) {
-      throw new Error(
-        `run_suspended ${eventId} was not assigned canonical coordinates`,
-      );
-    }
-    const suspension: AgenCBackgroundAgentSuspensionSnapshot = {
-      openedAt: active.startedAt,
-      epoch,
-      eventId,
-      sequence,
-      rolloutPath: active.bootstrap.rolloutStore.rolloutPath,
-      reason: pending.reason,
-      suspendedAt: pending.suspendedAt,
-    };
-    active.suspension = suspension;
-    try {
-      active.bootstrap.rolloutStore.recordRunSuspensionEvent(event);
-    } catch {
-      // Rebuildable projection; the fsync-committed event is authoritative.
-    }
-    return suspension;
-  } catch (error) {
-    const recovered = acceptCommitted(true);
-    if (recovered !== undefined) return recovered;
-    throw error;
-  }
-}
-
-function cancelledTerminalResult(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  stopReason: string,
-  finishedAt: string,
-): RunTerminalResult {
-  return {
-    runId,
-    status: "cancelled",
-    exitCode: null,
-    stopReason,
-    finalMessage: null,
-    usage: terminalUsageForActiveAgent(active),
-    lastSequence: null,
-    finishedAt,
-  };
-}
-
-function currentRunEpochFromRollout(
-  bootstrap: LocalRuntimeBootstrap,
-  runId: string,
-): number {
-  let epoch = 1;
-  try {
-    const items = bootstrap.rolloutStore.readAll() as ReadonlyArray<{
-      readonly type?: unknown;
-      readonly payload?: {
-        readonly msg?: {
-          readonly type?: unknown;
-          readonly payload?: {
-            readonly runId?: unknown;
-            readonly epoch?: unknown;
-          };
-        };
-      };
-    }>;
-    for (const item of items) {
-      if (
-        item.type !== "event_msg" ||
-        item.payload?.msg?.type !== "run_reopened" ||
-        item.payload.msg.payload?.runId !== runId
-      ) {
-        continue;
-      }
-      const reopenedEpoch = positiveSequence(item.payload.msg.payload.epoch);
-      if (reopenedEpoch !== undefined && reopenedEpoch > epoch) {
-        epoch = reopenedEpoch;
-      }
-    }
-  } catch {
-    // A new run has no reopen record. Read failures are surfaced when the
-    // terminal append itself tries to commit; epoch 1 is the only safe default.
-  }
-  return epoch;
-}
-
-function terminalResultFromThread(
-  active: ActiveBackgroundAgent,
-  runId: string,
-  status: TerminalThreadStatus,
-): RunTerminalResult {
-  const usage = terminalUsageForActiveAgent(active);
-  const finishedAt =
-    "endedAtMs" in status && Number.isFinite(status.endedAtMs)
-      ? new Date(status.endedAtMs).toISOString()
-      : active.lastActiveAt;
-  if (status.status === "completed") {
-    return {
-      runId,
-      status: "completed",
-      exitCode: 0,
-      stopReason: "turn_completed",
-      finalMessage: status.lastMessage ?? null,
-      usage,
-      lastSequence: null,
-      finishedAt,
-    };
-  }
-  if (status.status === "errored") {
-    return {
-      runId,
-      status: "failed",
-      exitCode: 1,
-      stopReason: status.error,
-      finalMessage: null,
-      usage,
-      lastSequence: null,
-      finishedAt,
-    };
-  }
-  return {
-    runId,
-    status: "cancelled",
-    exitCode: null,
-    stopReason: status.status === "shutdown" ? "shutdown" : "not_found",
-    finalMessage: null,
-    usage,
-    lastSequence: null,
-    finishedAt,
-  };
-}
-
-function mapThreadStatus(status: ThreadAgentStatus): DaemonAgentStatus {
-  switch (status.status) {
-    case "completed":
-    case "not_found":
-    case "shutdown":
-      return "stopped";
-    case "interrupted":
-    case "idle":
-      return "idle";
-    case "errored":
-      return "error";
-    case "pending_init":
-    case "running":
-      return "running";
-  }
-}
-
-function eventFromThreadStatus(
-  status: ThreadAgentStatus,
-): BackgroundAgentDaemonEvent | null {
-  switch (status.status) {
-    case "running":
-      return {
-        id: status.turnId,
-        type: "turn_started",
-        payload: {
-          turnId: status.turnId,
-          ...(status.startedAtMs !== undefined
-            ? { startedAt: status.startedAtMs }
-            : {}),
-        },
-      };
-    case "completed":
-      return {
-        id: status.turnId,
-        type: "turn_complete",
-        payload: {
-          turnId: status.turnId,
-          ...(status.lastMessage !== undefined
-            ? { lastAgentMessage: status.lastMessage }
-            : {}),
-          ...(status.endedAtMs !== undefined
-            ? { completedAt: status.endedAtMs }
-            : {}),
-        },
-      };
-    case "idle":
-      // Keep-alive worker between turns: surface a turn_complete so the daemon
-      // flips the agent to idle/completed instead of leaving it "running"
-      // forever. The id must be unique per transition (the keep-alive turnId is
-      // constant across turns), so suffix the monotonic end timestamp.
-      return {
-        id: `idle-${status.turnId}-${status.endedAtMs}`,
-        type: "turn_complete",
-        payload: {
-          turnId: status.turnId,
-          completedAt: status.endedAtMs,
-        },
-      };
-    case "errored":
-      return {
-        id: status.turnId,
-        type: "error",
-        payload: {
-          cause: "background_agent_error",
-          message: status.error,
-          turnId: status.turnId,
-        },
-      };
-    case "interrupted":
-      return {
-        id: `interrupted-${status.turnId}`,
-        type: "agent_status",
-        payload: {
-          status: "idle",
-          runStatus: "completed",
-          turnId: status.turnId,
-          message: status.reason,
-        },
-      };
-    case "shutdown":
-      return {
-        id: `shutdown-${status.endedAtMs}`,
-        type: "agent_status",
-        payload: {
-          status: "stopped",
-          runStatus: "stopped",
-          message: "shutdown",
-        },
-      };
-    case "not_found":
-      return {
-        id: "not-found",
-        type: "agent_status",
-        payload: {
-          status: "stopped",
-          runStatus: "stopped",
-          message: "not_found",
-        },
-      };
-    case "pending_init":
-      return null;
-  }
 }
 
 function eventFromProgress(
@@ -9793,16 +4809,6 @@ function messageText(content: unknown): string {
     .join("\n");
 }
 
-function messageContentToAgentInput(
-  content: MessageContent,
-): string | readonly LLMContentPart[] {
-  if (typeof content === "string") return content;
-  return content.map((part) => {
-    if (part.type === "text") return { type: "text", text: part.text };
-    return { type: "image_url", image_url: { url: part.image_url.url } };
-  });
-}
-
 function prepareDaemonUserPrompt(params: {
   readonly session: Session;
   readonly configStore: LocalRuntimeBootstrap["configStore"];
@@ -9818,208 +4824,6 @@ function prepareDaemonUserPrompt(params: {
         ? { hookPrompt: params.hookPrompt }
         : {}),
     }),
-  );
-}
-
-async function submitStructuredAgentInput(
-  active: ActiveBackgroundAgent,
-  input: readonly LLMContentPart[],
-  _displayText: string,
-  submitOptions?: SessionSubmitOptions,
-): Promise<void> {
-  try {
-    await active.thread.submit({
-      type: "user_input",
-      input,
-      ...(submitOptions !== undefined ? { submitOptions } : {}),
-    });
-  } catch (error) {
-    if (error instanceof MailboxClosedError) {
-      throw new Error(
-        `AgenC daemon agent not running: ${active.thread.threadId}`,
-      );
-    }
-    throw error;
-  }
-}
-
-function messageContentDisplayText(content: MessageContent): string {
-  if (typeof content === "string") return content;
-  return content
-    .map((part) => (part.type === "text" ? part.text : "[image]"))
-    .filter((part) => part.trim().length > 0)
-    .join("\n");
-}
-
-function interruptedToolResultContent(callId: string, reason: string): string {
-  return JSON.stringify({
-    tool_use_id: callId,
-    is_error: true,
-    content: `<tool_use_error>user interrupted - ${reason}</tool_use_error>`,
-  });
-}
-
-function hashStable(value: string): string {
-  let hash = 0;
-  for (let index = 0; index < value.length; index += 1) {
-    hash = (hash * 31 + value.charCodeAt(index)) >>> 0;
-  }
-  return hash.toString(36);
-}
-
-export function resolvePermissionDecisionTimeoutMs(
-  env: NodeJS.ProcessEnv = process.env,
-): number | undefined {
-  const raw = env.AGENC_PERMISSION_TIMEOUT_MS;
-  if (raw === undefined) return undefined;
-  const parsed = Number.parseInt(raw, 10);
-  return Number.isFinite(parsed) && parsed > 0 ? parsed : undefined;
-}
-
-function readApprovalAgentId(ctx: ApprovalCtx): string | null {
-  const session = ctx.invocation.session as { conversationId?: unknown };
-  return typeof session.conversationId === "string" &&
-    session.conversationId.length > 0
-    ? session.conversationId
-    : null;
-}
-
-/**
- * For an ExitPlanMode approval, enrich the request_permissions payload with the
- * plan content and path so the TUI overlay can render the plan being approved.
- * Falls back to the tool input's `plan` string when the on-disk plan is empty.
- * Returns an empty object for any other tool.
- *
- * Exported (under a test-scoped name) so the enrichment can be unit-tested
- * directly with mocked getPlan/getPlanFilePath without bootstrapping an agent.
- */
-export function planApprovalPayloadFields(
-  toolName: string,
-  agentId: string,
-  input: JsonObject,
-): JsonObject {
-  if (toolName !== EXIT_PLAN_MODE_TOOL_NAME) return {};
-  const fields: Record<string, JsonValue> = {};
-  const agent = agentId as AgentId;
-  let planContent: string | null = null;
-  try {
-    planContent = getPlan(agent);
-  } catch {
-    planContent = null;
-  }
-  if (
-    (planContent === null || planContent.length === 0) &&
-    typeof input.plan === "string" &&
-    input.plan.length > 0
-  ) {
-    planContent = input.plan;
-  }
-  if (typeof planContent === "string" && planContent.length > 0) {
-    fields.planContent = planContent;
-  }
-  let planFilePath: string | undefined;
-  try {
-    planFilePath = getPlanFilePath(agent);
-  } catch {
-    planFilePath = undefined;
-  }
-  if (typeof planFilePath === "string" && planFilePath.length > 0) {
-    fields.planFilePath = planFilePath;
-  }
-  return fields;
-}
-
-function restoreBootstrapSelection(params: AgenCBackgroundAgentRestoreParams): {
-  readonly provider?: string;
-  readonly model?: string;
-  readonly profile?: string;
-  readonly configPath?: string;
-  readonly addDirs?: readonly string[];
-  readonly permissionMode?:
-    | "default"
-    | "plan"
-    | "acceptEdits"
-    | "bypassPermissions"
-    | "dontAsk"
-    | "auto";
-} {
-  const canonical = params.runtimeSettings;
-  if (canonical === undefined) return params;
-  return {
-    provider: canonical.provider,
-    model: canonical.model,
-    ...(canonical.profile !== null ? { profile: canonical.profile } : {}),
-    ...(params.configPath !== undefined
-      ? { configPath: params.configPath }
-      : {}),
-    ...(params.addDirs !== undefined ? { addDirs: params.addDirs } : {}),
-  };
-}
-
-function runtimeSettingsWithRestoreOverrides(
-  canonical: RunRuntimeSettingsSnapshot,
-  params: AgenCBackgroundAgentRestoreParams,
-  workspaceRoot: string,
-  config: AgenCConfig,
-): RunRuntimeSettingsSnapshot {
-  const permissionMode = params.permissionMode ?? canonical.permissionMode;
-  const permissionChanged = permissionMode !== canonical.permissionMode;
-  const prePlanMode =
-    permissionMode === "plan"
-      ? permissionChanged
-        ? canonical.permissionMode
-        : canonical.prePlanMode
-      : null;
-  const bypassTransitionCritical =
-    permissionMode === "bypassPermissions" ||
-    prePlanMode === "bypassPermissions";
-  const resolvedSelection = resolveProviderModelSelection(
-    config,
-    { provider: canonical.provider, model: canonical.model },
-    {
-      ...(params.provider !== undefined
-        ? { model_provider: params.provider }
-        : {}),
-      ...(params.model !== undefined ? { model: params.model } : {}),
-    },
-  );
-  return {
-    ...canonical,
-    permissionMode,
-    prePlanMode,
-    autoModeActive:
-      permissionMode === "auto"
-        ? true
-        : permissionMode === "plan" && !permissionChanged
-          ? canonical.autoModeActive
-          : false,
-    bypassPermissionsWorkspace: bypassTransitionCritical ? workspaceRoot : null,
-    model: resolvedSelection.model,
-    provider: resolvedSelection.provider,
-    profile: params.profile ?? canonical.profile,
-  };
-}
-
-function buildBootstrapArgv(
-  params: {
-    readonly provider?: string;
-    readonly model?: string;
-    readonly profile?: string;
-    readonly configPath?: string;
-    readonly addDirs?: readonly string[];
-    readonly permissionMode?:
-      | "default"
-      | "plan"
-      | "acceptEdits"
-      | "bypassPermissions"
-      | "dontAsk"
-      | "auto";
-  },
-  executableArgv: readonly string[] | undefined,
-): readonly string[] {
-  return buildStructuredSessionBootstrapArgv(
-    params,
-    executableArgv ?? [process.execPath, process.argv[1] ?? "agenc"],
   );
 }
 
@@ -10122,20 +4926,3 @@ function installDaemonTurnDriverHooks(
 
 export const __installDaemonTurnDriverHooksForTest =
   installDaemonTurnDriverHooks;
-
-async function installUnattendedPermissionPolicy(
-  registry: PermissionModeRegistry,
-  allow: readonly string[] | undefined,
-  deny: readonly string[] | undefined,
-): Promise<void> {
-  const next = applyUnattendedPermissionPolicyToContext(registry.current(), {
-    ...(allow !== undefined ? { allowlist: allow } : {}),
-    ...(deny !== undefined ? { denylist: deny } : {}),
-  });
-  await registry.update(next);
-}
-
-// runRestoredAgentToCompletion / restoredAgentMetadata are retained for
-// compatibility with the older fork-loop restore path while the live
-// ManagedThread restore path handles replay directly above.
-void [runRestoredAgentToCompletion, restoredAgentMetadata];
